### 2025-12-23 d9d2a0f
remove draft paper workflow for now
