### 2025-12-24 #738 @mpadge
@smwindecker Thanks for your submission which inadvertently help reveal a :bug: in our current deploy. The bot should have replied with this message:

> ⭐️ Thanks for submitting to rOpenSci.  We are taking a break for the holidays. Activities are currently paused until Jan 5th 2026. Please alert us here on or after that date to proceed with your submission. ⭐️

Note also that stats submissions generally start with a pre-submission inquiry before moving on to the full submission. But your package is obviously in great shape already, so once we get moving we'll be able to proceed within this current full issue.
