+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/algorithmic-scaling: algorithmic-scaling.html
   articles/parameter-recovery: parameter-recovery.html
   Using-EpiStrainDynamics: Using-EpiStrainDynamics.html
-last_built: 2025-12-23T00:24Z
+last_built: 2026-01-05T10:27Z
 urls:
   reference: https://acefa-hubs.github.io/EpiStrainDynamics/reference
   article: https://acefa-hubs.github.io/EpiStrainDynamics/articles
