+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/algorithmic-scaling: algorithmic-scaling.html
   articles/parameter-recovery: parameter-recovery.html
   Using-EpiStrainDynamics: Using-EpiStrainDynamics.html
-last_built: 2026-01-05T10:27Z
+last_built: 2026-01-06T00:56Z
 urls:
   reference: https://acefa-hubs.github.io/EpiStrainDynamics/reference
   article: https://acefa-hubs.github.io/EpiStrainDynamics/articles
