### 2026-01-08 0f8516b
ensure ingest functions can take timeseries objects
