+++ R/validation-utils.R
@@ -233,18 +233,34 @@ validate_smoothing_structure <- function(smoothing_obj, pathogen_names = NULL,
 
 # Helper: Validate column exists in data
 #' @noRd
-check_column_exists <- function(data, col_name, arg_name) {
-  if (!col_name %in% names(data)) {
-    stop(sprintf("Column '%s' specified in %s not found in data", col_name, arg_name))
+check_column_exists <- function(data, col_name) {
+  if (!col_name %in% colnames(data)) {
+    stop(sprintf("Column '%s' not found in data", col_name))
   }
 }
 
-# Helper: Validate column is numeric
+#' Helper: Detect if data is a time series class
+#'
+#' @param data input data object
+#'
+#' @returns logical indicating if data is a time series class
 #' @noRd
-check_column_numeric <- function(data, col_name, arg_name) {
-  if (!is.numeric(data[[col_name]])) {
-    stop(sprintf("Column '%s' specified in %s must be numeric, but is %s",
-                 col_name, arg_name, class(data[[col_name]])[1]))
+#' @srrstats {G1.4} uses `Roxygen2` documentation
+#' @srrstats {G1.4a} internal function specified with `@noRd`
+is_timeseries_class <- function(data) {
+  ts_classes <- c("ts", "mts", "xts", "zoo", "zooreg", "tsibble", "tbl_ts",
+                  "tibbletime", "tbl_time", "timeSeries", "irts", "tis")
+  any(class(data) %in% ts_classes)
+}
+
+# Helper: Validate column is numeric
+#' @srrstats {TS1.7} accommodate potential units data input
+#' @noRd
+check_column_numeric <- function(data, col_name) {
+  if (!is.numeric(data[[col_name]]) && !inherits(data[[col_name]], "units")) {
+    stop(sprintf("Column '%s' specified must be numeric or have units,
+                 but is %s",
+                 col_name, class(data[[col_name]])[1]))
   }
 }
 
@@ -291,48 +307,193 @@ check_missing_data <- function(data, columns, context) {
   invisible(NULL)
 }
 
-#' Helper: Create and validate tsibble from dataframe
+#' Helper: Convert time series objects to tsibble
 #'
-#' @param data input dataset to check time formatting
-#' @param columns column names with data to include in the validated df
-#' @param time_col the column name with time data to validate
+#' @param ts_obj time series object (ts, mts, xts, zoo, etc.)
+#'
+#' @importFrom timetk tk_tbl
+#'
+#' @returns tsibble object in wide format
+#' @noRd
+#' @srrstats {G1.4} uses `Roxygen2` documentation
+#' @srrstats {G1.4a} internal function specified with `@noRd`
+convert_ts_to_tsibble <- function(ts_obj) {
+
+  # Handle xts and zoo objects using timetk (best practice for these classes)
+  if (inherits(ts_obj, c("xts", "zoo", "zooreg"))) {
+
+    # Check if timetk is available
+    if (!requireNamespace("timetk", quietly = TRUE)) {
+      cli::cli_abort(
+        c(
+          "The {.pkg timetk} package is required to handle
+          {.cls {class(ts_obj)[1]}} objects.",
+          "i" = "Install it with: install.packages('timetk')"
+        )
+      )
+    }
+
+    # Convert xts/zoo to tibble using timetk (preserves wide format)
+    temp_tbl <- timetk::tk_tbl(ts_obj, preserve_index = TRUE, rename_index = "index")
+
+    # Convert to tsibble
+    temp_tsbl <- tryCatch({
+      tsibble::as_tsibble(temp_tbl, index = .data$index)
+    }, error = function(e) {
+      cli::cli_abort(
[…]
