+++ tests/testthat/test-timeseries-helpers.R
@@ -1,0 +1,481 @@
+# Tests for timeseries helper functions
+# These tests focus on validation checks and edge cases in the helper functions
+
+# ==============================================================================
+# Tests for is_timeseries_class()
+# ==============================================================================
+
+test_that("is_timeseries_class() correctly identifies time series objects", {
+  # Test with data.frame (should be FALSE)
+  df <- data.frame(x = 1:10, y = 1:10)
+  expect_false(is_timeseries_class(df))
+
+  # Test with tibble (should be FALSE)
+  if (requireNamespace("tibble", quietly = TRUE)) {
+    tbl <- tibble::tibble(x = 1:10, y = 1:10)
+    expect_false(is_timeseries_class(tbl))
+  }
+
+  # Test with ts (should be TRUE)
+  ts_obj <- ts(1:10, start = 1, frequency = 1)
+  expect_true(is_timeseries_class(ts_obj))
+
+  # Test with mts (should be TRUE)
+  mts_obj <- ts(matrix(1:20, ncol = 2), start = 1, frequency = 1)
+  expect_true(is_timeseries_class(mts_obj))
+
+  # Test with xts (should be TRUE)
+  if (requireNamespace("xts", quietly = TRUE)) {
+    xts_obj <- xts::xts(1:10, order.by = Sys.Date() + 0:9)
+    expect_true(is_timeseries_class(xts_obj))
+  }
+
+  # Test with zoo (should be TRUE)
+  if (requireNamespace("zoo", quietly = TRUE)) {
+    zoo_obj <- zoo::zoo(1:10, order.by = 1:10)
+    expect_true(is_timeseries_class(zoo_obj))
+  }
+
+  # Test with tsibble (should be TRUE)
+  if (requireNamespace("tsibble", quietly = TRUE)) {
+    tsbl_obj <- tsibble::tsibble(
+      time = 1:10,
+      value = 1:10,
+      index = time
+    )
+    expect_true(is_timeseries_class(tsbl_obj))
+  }
+})
+
+# ==============================================================================
+# Tests for convert_ts_to_tsibble()
+# ==============================================================================
+
+test_that("convert_ts_to_tsibble() handles xts objects correctly", {
+  skip_if_not_installed("xts")
+  skip_if_not_installed("timetk")
+
+  # Create xts object
+  dates <- seq.Date(as.Date("2020-01-01"), by = "day", length.out = 10)
+  xts_obj <- xts::xts(
+    data.frame(a = 1:10, b = 11:20),
+    order.by = dates
+  )
+
+  result <- convert_ts_to_tsibble(xts_obj)
+
+  expect_s3_class(result, "tbl_ts")
+  expect_true("index" %in% names(result))
+  expect_equal(nrow(result), 10)
+})
+
+test_that("convert_ts_to_tsibble() handles zoo objects correctly", {
+  skip_if_not_installed("zoo")
+  skip_if_not_installed("timetk")
+
+  # Create zoo object
+  dates <- seq.Date(as.Date("2020-01-01"), by = "day", length.out = 10)
+  zoo_obj <- zoo::zoo(
+    data.frame(a = 1:10, b = 11:20),
+    order.by = dates
+  )
+
+  result <- convert_ts_to_tsibble(zoo_obj)
+
+  expect_s3_class(result, "tbl_ts")
+  expect_true("index" %in% names(result))
+  expect_equal(nrow(result), 10)
+})
+
+test_that("convert_ts_to_tsibble() keeps multivariate ts in wide format", {
+  # Create multivariate time series
+  mts_obj <- ts(matrix(1:30, ncol = 3), start = 1, frequency = 1)
[…]
