+++ NAMESPACE
@@ -55,6 +55,7 @@ importFrom(splines,bs)
 importFrom(stats,predict)
 importFrom(stats,quantile)
 importFrom(stats,setNames)
+importFrom(timetk,tk_tbl)
 importFrom(utils,head)
 importFrom(viridis,viridis)
 useDynLib(EpiStrainDynamics, .registration = TRUE)

[…]
