### 2026-01-08 720a515
ensure ingest functions can take timeseries objects
