+++ tests/testthat/test-pathogen-structures.R
@@ -25,13 +25,13 @@ test_that("single() validates column existence and types", {
   # Test non-existent time column
   expect_error(
     single(data = sarscov2, case_timeseries = 'cases', time = 'nonexistent'),
-    "time"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent case column
   expect_error(
     single(data = sarscov2, case_timeseries = 'nonexistent', time = 'date'),
-    "case_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-numeric case column (if you have a non-numeric column)
@@ -99,7 +99,7 @@ test_that("multiple() validates column existence and types", {
       time = 'nonexistent',
       component_pathogen_timeseries = c('alpha')
     ),
-    "time"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent case column
@@ -110,7 +110,7 @@ test_that("multiple() validates column existence and types", {
       time = 'date',
       component_pathogen_timeseries = c('alpha')
     ),
-    "case_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent component pathogen column
@@ -121,7 +121,7 @@ test_that("multiple() validates column existence and types", {
       time = 'date',
       component_pathogen_timeseries = c('alpha', 'nonexistent')
     ),
-    "component_pathogen_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-numeric component pathogen column
@@ -202,7 +202,7 @@ test_that("subtyped() validates column existence and types", {
       influenzaA_subtyped_timeseries = c('inf_H3N2'),
       other_pathogen_timeseries = c('inf_B')
     ),
-    "time"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent case column
@@ -215,7 +215,7 @@ test_that("subtyped() validates column existence and types", {
       influenzaA_subtyped_timeseries = c('inf_H3N2'),
       other_pathogen_timeseries = c('inf_B')
     ),
-    "case_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent influenzaA_unsubtyped column
@@ -228,7 +228,7 @@ test_that("subtyped() validates column existence and types", {
       influenzaA_subtyped_timeseries = c('inf_H3N2'),
       other_pathogen_timeseries = c('inf_B')
     ),
-    "influenzaA_unsubtyped_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent influenzaA_subtyped column
@@ -241,7 +241,7 @@ test_that("subtyped() validates column existence and types", {
       influenzaA_subtyped_timeseries = c('nonexistent'),
       other_pathogen_timeseries = c('inf_B')
     ),
-    "influenzaA_subtyped_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-existent other_pathogen column
@@ -254,7 +254,7 @@ test_that("subtyped() validates column existence and types", {
       influenzaA_subtyped_timeseries = c('inf_H3N2'),
       other_pathogen_timeseries = c('nonexistent')
     ),
-    "other_pathogen_timeseries"
+    "Column 'nonexistent' not found in data"
   )
 
   # Test non-numeric columns
