### 2026-01-12 #738 @smwindecker
Thanks @mpadge!

re large files: I have removed the large files from the history. I may have gotten a bit overzealous and removed some files that should be there, but so far I can't identify any impacts.

re timeseries: Yes the package ingests timeseries data. After initially selecting only the bayesian srr standards, I only began to consider the impact/need to consider timeseries formats when I addressed standards G2.15, BS3.0, G5.3 regarding whether data assumes and returns non-missingness. To comply with these standards I implemented a validation step which relied on tsibble functions to check for gaps, regularity, and ordering of data. Based on your advice I have revisited those standards. As there is no forecasting most of them are in the NA pile, the others I have addressed formally using the srr tags in the new version. I am running the final rebuild checks now will make a note when they have all passed, hopefully later today.

Thank you for the kind invitation to contribute the two articles, I really appreciate the positive feedback!
