### 2026-01-12 038ff00
1. ingest timeseries relevant data 2. return timeseries formatted data
