### 2026-01-12 35b7ac3
1. ingest timeseries relevant data 2. return timeseries formatted data
