### 2026-01-12 c5d3c5b
setup incidence function so it has a default for dow that matches the dow in the model.
