### 2026-01-12 de19ad0
setup incidence function so it has a default for dow that matches the dow in the model.
