+++ tests/testthat/test-metrics.R
@@ -27,6 +27,7 @@ test_that("fixture models are available", {
 
 # Load fixtures (done once per test file)
 fit_rw_single <- readRDS("tests/testthat/fixtures/fit_rw_single.rds")
+fit_rw_single_dow <- readRDS("tests/testthat/fixtures/fit_rw_single_dow.rds")
 fit_ps_single <- readRDS("tests/testthat/fixtures/fit_ps_single.rds")
 fit_rw_multi <- readRDS("tests/testthat/fixtures/fit_rw_multi.rds")
 fit_ps_multi <- readRDS("tests/testthat/fixtures/fit_ps_multi.rds")
@@ -162,8 +163,8 @@ test_that("Rt() threshold affects prop calculation", {
 # ==============================================================================
 # TESTS: incidence() FUNCTION
 # ==============================================================================
-
 test_that("incidence() validates inputs correctly", {
+  # Invalid class
   expect_error(
     incidence(list(fit = "not a fit"), dow = FALSE),
     "Input must be of class"
@@ -172,10 +173,57 @@ test_that("incidence() validates inputs correctly", {
   # dow = TRUE when model doesn't have dow_effect
   expect_error(
     incidence(fit_rw_single, dow = TRUE),
-    "dow effects can't be incorporated"
+    "Day-of-week effects cannot be incorporated"
+  )
+
+  # dow must be NULL or logical
+  expect_error(
+    incidence(fit_rw_single, dow = "yes"),
+    "must be.*NULL.*logical"
+  )
+
+  expect_error(
+    incidence(fit_rw_single, dow = 1),
+    "must be.*NULL.*logical"
   )
 })
 
+# dow argument default behavior tests
+test_that("incidence() uses model's dow setting when dow = NULL", {
+  # For model without dow_effect
+  expect_message(
+    result <- incidence(fit_rw_single, dow = NULL),
+    "Using day-of-week setting from model.*FALSE"
+  )
+  expect_s3_class(result, "incidence")
+})
+
+test_that("incidence() default behavior (dow not specified) uses model setting", {
+  # When dow argument is not provided at all, should use model's setting
+  expect_message(
+    result <- incidence(fit_rw_single),  # dow not specified
+    "Using day-of-week setting from model"
+  )
+  expect_s3_class(result, "incidence")
+})
+
+test_that("incidence() explicit dow = FALSE works regardless of model", {
+  # Should work for model without dow_effect
+  expect_silent(
+    result <- incidence(fit_rw_single, dow = FALSE)
+  )
+  expect_s3_class(result, "incidence")
+})
+
+test_that("incidence() dow = TRUE requires model to have dow_effect", {
+  # Model without dow_effect - should error
+  expect_error(
+    incidence(fit_rw_single, dow = TRUE),
+    "Day-of-week effects cannot be incorporated"
+  )
+})
+
+# Basic functionality tests (existing tests - reviewed and kept)
 test_that("incidence() works for single pathogen models", {
   result <- incidence(fit_rw_single, dow = FALSE)
 
@@ -215,7 +263,6 @@ test_that("incidence() works for multiple pathogen models", {
   # Total should be sum of individual pathogens (approximately, given posterior sampling)
   total_data <- result$measure[result$measure$pathogen == "Total", ]
   alpha_data <- result$measure[result$measure$pathogen == "alpha", ]
-
   expect_equal(nrow(total_data), nrow(alpha_data))
 })
 
@@ -228,6 +275,154 @@ test_that("incidence() p-spline vs random walk gives different results", {
 
   # But values should differ (different smoothing methods)
   expect_false(all(abs(result_rw$measure$y - result_ps$measure$y) < 1e-10))
+})
+
[…]
