### 2026-01-12 9da9785
add descriptions of the timeseries class ingestion
