### 2026-01-13 7d7e724
remove NA standard
