### 2026-01-13 f43669b
remove NA standard
