### 2026-01-13 f87397d
revise typo of input argument name
