### 2026-01-13 0195eb5
deal with unquoted operator
