### 2026-01-13 61b61cf
update fit_model documentation
