### 2026-01-13 f8f274a
update tests with new data lists in outputs
