### 2026-01-13 6115aae
update tests with new data lists in outputs
