+++ tests/testthat/test-metrics.R
@@ -9,28 +9,28 @@ test_that("fixture models are available", {
   # These should be created once and saved as .rds files
   # See setup-fixtures.R for generation code
 
-  skip_if_not(file.exists("tests/testthat/fixtures/fit_rw_single.rds"),
+  skip_if_not(file.exists(test_path("fixtures/fit_rw_single.rds")),
               "Fixture fit_rw_single.rds not found")
-  skip_if_not(file.exists("tests/testthat/fixtures/fit_ps_single.rds"),
+  skip_if_not(file.exists(test_path("fixtures/fit_ps_single.rds")),
               "Fixture fit_ps_single.rds not found")
-  skip_if_not(file.exists("tests/testthat/fixtures/fit_rw_multi.rds"),
+  skip_if_not(file.exists(test_path("fixtures/fit_rw_multi.rds")),
               "Fixture fit_rw_multi.rds not found")
-  skip_if_not(file.exists("tests/testthat/fixtures/fit_ps_multi.rds"),
+  skip_if_not(file.exists(test_path("fixtures/fit_ps_multi.rds")),
               "Fixture fit_ps_multi.rds not found")
 
   # If we get here, all fixtures exist
-  expect_true(file.exists("tests/testthat/fixtures/fit_rw_single.rds"))
-  expect_true(file.exists("tests/testthat/fixtures/fit_ps_single.rds"))
-  expect_true(file.exists("tests/testthat/fixtures/fit_rw_multi.rds"))
-  expect_true(file.exists("tests/testthat/fixtures/fit_ps_multi.rds"))
+  expect_true(file.exists(test_path("fixtures/fit_rw_single.rds")))
+  expect_true(file.exists(test_path("fixtures/fit_ps_single.rds")))
+  expect_true(file.exists(test_path("fixtures/fit_rw_multi.rds")))
+  expect_true(file.exists(test_path("fixtures/fit_ps_multi.rds")))
 })
 
 # Load fixtures (done once per test file)
-fit_rw_single <- readRDS("tests/testthat/fixtures/fit_rw_single.rds")
-fit_rw_single_dow <- readRDS("tests/testthat/fixtures/fit_rw_single_dow.rds")
-fit_ps_single <- readRDS("tests/testthat/fixtures/fit_ps_single.rds")
-fit_rw_multi <- readRDS("tests/testthat/fixtures/fit_rw_multi.rds")
-fit_ps_multi <- readRDS("tests/testthat/fixtures/fit_ps_multi.rds")
+fit_rw_single <- readRDS(test_path("fixtures/fit_rw_single.rds"))
+fit_rw_single_dow <- readRDS(test_path("fixtures/fit_rw_single_dow.rds"))
+fit_ps_single <- readRDS(test_path("fixtures/fit_ps_single.rds"))
+fit_rw_multi <- readRDS(test_path("fixtures/fit_rw_multi.rds"))
+fit_ps_multi <- readRDS(test_path("fixtures/fit_ps_multi.rds"))
 
 # Helper function for generation interval
 gi_simple <- function(x) {
@@ -83,11 +83,11 @@ test_that("Rt() works for single pathogen random walk model", {
   # Check measure data frame
   expect_s3_class(result$measure, "data.frame")
   expect_named(result$measure,
-               c("y", "lb_50", "ub_50", "lb_95",
-                 "ub_95", "prop", "time", "pathogen"))
+               c("time", "y", "lb_50", "ub_50", "lb_95",
+                 "ub_95", "prop", "pathogen"))
 
   # Check dimensions - should start from time index = tau_max
-  n_days <- length(fit_rw_single$constructed_model$data$time)
[…]
