+++ tests/testthat/test-fit_model.R
@@ -341,47 +341,3 @@ test_that("verbose = FALSE suppresses Stan output", {
   # Should have minimal/no output
   expect_lt(length(output), 5)
 })
-
-# ==============================================================================
-# OPTIONAL: REGRESSION TEST WITH FIXTURE (if desired)
-# ==============================================================================
-
-test_that("fit_model produces consistent results with fixture", {
-  # This uses the fixture from test-metrics.R if available
-  skip_if_not(file.exists("tests/testthat/fixtures/fit_rw_single.rds"),
-              "Fixture not available")
-
-  # Load the fixture
-  fit_fixture <- readRDS("tests/testthat/fixtures/fit_rw_single.rds")
-
-  # Refit with same data and seed
-  test_data <- fit_fixture$constructed_model$data
-  mod <- construct_model(
-    method = random_walk(),
-    pathogen_structure = single(
-      data = data.frame(
-        dates = test_data$time,
-        cases = test_data$case_timeseries
-      ),
-      case_timeseries = 'cases',
-      time = 'dates'
-    )
-  )
-
-  suppressWarnings(
-    fit_new <- fit_model(mod, n_iter = 1000, n_chain = 2,
-                         verbose = FALSE, seed = 42)
-  )
-
-  # Check structure is the same
-  expect_s3_class(fit_new, "EpiStrainDynamics.fit")
-  expect_s3_class(fit_new, "rw_single")
-
-  # Note: We don't expect exact numerical equivalence due to MCMC stochasticity
-  # even with the same seed (Stan version differences, etc.)
-  # But we can check dimensions are consistent
-  post_fixture <- rstan::extract(fit_fixture$fit)
-  post_new <- rstan::extract(fit_new$fit)
-
-  expect_equal(dim(post_fixture$a), dim(post_new$a))
-})
