### 2026-01-13 e05cfd3
ensure Rplots.pdf not created
