### 2026-01-13 #738 @mpadge
Here are some ideas for how you could extend compliance:

## General standards

The following standards were written to help ensure software is compatible with piping-type workflows, in which things like list-columns may inadvertently appear, and missing data may not necessarily be known in advance:

- [**G2.11**](https://stats-devguide.ropensci.org/standards.html#:~:text=G2.11) is already complied with, as main functions appear to accept vector columns with attributes.
- [**G2.12**](https://stats-devguide.ropensci.org/standards.html#:~:text=G2.12) should be complied with. Function currently errors if `case_timeseries` is a list-column (`sarscov2$cases <- I(as.list(sarscov2$cases))`), but should "just work".
- Complying with [**G2.14**](https://stats-devguide.ropensci.org/standards.html#:~:text=G2.14) would make the package more robust. Currently any missing values generate a (very informative) error. Implementing other options would help general workflows for which missingness may not necessarily be known in advance.

Additionally:

- [**G2.6**](https://stats-devguide.ropensci.org/standards.html#:~:text=G2.6) is intended to apply to your parameters like `component_pathogen_timeseries`, `tau_mean`, and `tau_sd`. Directly complying with this would like improve pre-processing and checks on those kinds of parameters.
- [**G5.0**](https://stats-devguide.ropensci.org/standards.html#:~:text=G5.0) is already [complied with](https://acefa-epi-analytics.org/EpiStrainDynamics/reference/sarscov2.html).
- [**G5.6**](https://stats-devguide.ropensci.org/standards.html#:~:text=G5.6) is clearly [complied with](https://acefa-epi-analytics.org/EpiStrainDynamics/articles/parameter-recovery.html).

And there are likely to be others with which you could comply too.

## Bayesian standards

- [**BS2.13**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS2.13) and [**BS2.14**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS2.14) could likely easily be complied with.
- [**BS2.6**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS2.6) is complied with; just change from `srrstatsNA` to `srrstats`.
- [**BS4.6**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS4.6),  [**BS4.7**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS4.7),  [**BS5.3**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS5.3), and [**BS5.4**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS5.4) are probably already complied with via stan, which is perfectly fine. A bit of info on how stan meets those might be needed to elevate from `srrstatsNA` to `srrstats`.
- [**BS2.13**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS2.13) and [**BS2.14**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS2.14) are likely complied with, but just might need a bit of checking out how stan controls those. Don't worry about separation of progress bars from other stuff; main point is verbosity of warnings versus other stuff.
- [**BS6.0**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS6.0) is already complied with.
- [**BS6.1**](https://stats-devguide.ropensci.org/standards.html#standards-bayesian#:~:text=BS6.1) probably is too.

... and probably several more, but hopefully you get the idea. Generally, if your package complies indirectly through stan, then it complies, and statements should just indicate that.

---

While you're working on those, you can use [`srr::srr_stats_pre_submit()`](https://docs.ropensci.org/srr/reference/srr_stats_pre_submit.html) to check and confirm your compliance.
