### 2026-02-04 db0dd4f
add relevant srr given a convergence checker has been implemented
