### 2026-02-04 1192750
add suppress_warnings argument to allow user more detailed control over which messages to print
wrap stan call in a tryCatch
