+++ R/fit_model.R
@@ -8,13 +8,18 @@
 #' @param n_iter A positive integer specifying the number of iterations for each chain, default value is  2000
 #' @param n_warmup A positive integer specifying the number of warmup iterations, default value is half the number of iterations
 #' @param thin A positive integer specifying the period for saving samples, default value is 1.
-#' @param adapt_delta Numeric value between 0 and 1 indicating target average acceptance probability used in `rstan::sampling`. Default value is 0.8.
+#' @param adapt_delta Numeric value between 0 and 1 indicating target average acceptance probability used in `rstan::sampling`. Default value is 0.9.
 #' @param multi_cores A logical value indicating whether to parallelize chains with multiple cores, default is TRUE and uses all available cores - 1.
-#' @param verbose Logical value controlling the verbosity of output (i.e., warnings, messages, progress bar), default is TRUE.
+#' @param verbose Logical value controlling the verbosity of output. When TRUE (default),
+#'   shows all messages, warnings, errors, and progress indicators. When FALSE, suppresses
+#'   messages and progress while retaining warnings and errors.
+#' @param suppress_warnings Logical value indicating whether to suppress warnings from Stan.
+#'   Default is FALSE. When TRUE, warnings are suppressed but errors are still raised.
 #' @param seed A positive integer seed used for random number generation in MCMC. Default is NULL, which means the seed is generated from 1 to the maximum integer supported by R.
 #' @param ... additional arguments to `rstan::sampling()`, such as `init`
 #'
-#' @returns fit model of class `EpiStrainDynamics.fit`
+#' @returns fit model of class `EpiStrainDynamics.fit`, or if fitting fails,
+#'   an error is raised that can be caught and inspected.
 #' @export
 #'
 #' @srrstats {G1.4} uses `Roxygen2` documentation
@@ -23,14 +28,25 @@
 #'   in the documentation.
 #' @srrstats {BS2.7, BS2.11} Starting values can be specified with the
 #'   `init` argument to `rstan::sampling()` specified optionally here.
-#' @srrstats {BS2.12} Argument `verbose`, which defaults to `TRUE` controls the
-#'   verbosity of the stan sampling output.
-#' @srrstats {BS2.15} checks that data has been input in correct form, and
-#'   provides an informative error message if not
+#' @srrstats {BS2.12} Argument `verbose`, which defaults to `TRUE`, controls the
+#'   verbosity of output, showing all messages, warnings, errors, and progress
+#'   indicators.
+#' @srrstats {BS2.13} When `verbose = FALSE`, messages and progress indicators
+#'   are suppressed (via `refresh = 0` and `show_messages = FALSE`) while
+#'   warnings and errors are retained.
+#' @srrstats {BS2.14} Argument `suppress_warnings`, which defaults to `FALSE`,
+#'   enables suppression of warnings when set to `TRUE`.
+#' @srrstats {BS2.15} Errors from Stan sampling are caught via `tryCatch` and
+#'   re-raised with additional context. Stan failures (e.g., initialization
+#'   failures) that don't throw errors are detected and converted to errors.
+#'   The error object contains the original error message and the constructed
+#'   model for inspection.
 #' @srrstats {BS5.1} Returned list is of class `EpiStrainDynamics.fit`, and
 #'   includes the fit object of class `stanfit` and the pre-specified
 #'   constructed model object of class `EpiStrainDynamics.model`. The
[…]
