### 2026-02-04 e14bf2e
add relevant srr for additional standard governing tabular data
