### 2026-02-04 9179a5c
add check for list columns
move mcmc validation functions to this file
