+++ R/validate_mcmc_parameters.R
@@ -1,271 +1,0 @@
-#' Validate MCMC Chain Parameter
-#'
-#' Validates that n_chain is a positive integer
-#'
-#' @param n_chain The number of MCMC chains
-#'
-#' @noRd
-#' @srrstats {G1.4} uses `Roxygen2` documentation
-#' @srrstats {G1.4a} internal function specified with `@noRd`
-#' @srrstats {G5.2a} every error statement is unique
-#'
-#' @return NULL (invisibly) if validation passes, otherwise stops with error
-validate_n_chain <- function(n_chain) {
-  if (!is.numeric(n_chain)) {
-    cli::cli_abort("{.arg n_chain} must be numeric, got {.cls {class(n_chain)}}")
-  }
-
-  if (length(n_chain) != 1) {
-    cli::cli_abort("{.arg n_chain} must be a single value, got length {length(n_chain)}")
-  }
-
-  if (!is.finite(n_chain)) {
-    cli::cli_abort("{.arg n_chain} must be a finite number")
-  }
-
-  if (n_chain != as.integer(n_chain)) {
-    cli::cli_abort("{.arg n_chain} must be a whole number, got {n_chain}")
-  }
-
-  if (n_chain <= 0) {
-    cli::cli_abort("{.arg n_chain} must be positive, got {n_chain}")
-  }
-
-  invisible(NULL)
-}
-
-#' Validate MCMC Iteration Parameter
-#'
-#' Validates that n_iter is a positive integer
-#'
-#' @param n_iter The number of MCMC iterations
-#'
-#' @noRd
-#' @srrstats {G1.4} uses `Roxygen2` documentation
-#' @srrstats {G1.4a} internal function specified with `@noRd`
-#' @srrstats {G5.2a} every error statement is unique
-#'
-#' @return NULL (invisibly) if validation passes, otherwise stops with error
-validate_n_iter <- function(n_iter) {
-  if (!is.numeric(n_iter)) {
-    cli::cli_abort("{.arg n_iter} must be numeric, got {.cls {class(n_iter)}}")
-  }
-
-  if (length(n_iter) != 1) {
-    cli::cli_abort("{.arg n_iter} must be a single value, got length {length(n_iter)}")
-  }
-
-  if (!is.finite(n_iter)) {
-    cli::cli_abort("{.arg n_iter} must be a finite number")
-  }
-
-  if (n_iter != as.integer(n_iter)) {
-    cli::cli_abort("{.arg n_iter} must be a whole number, got {n_iter}")
-  }
-
-  if (n_iter <= 0) {
-    cli::cli_abort("{.arg n_iter} must be positive, got {n_iter}")
-  }
-
-  invisible(NULL)
-}
-
-#' Validate MCMC Warmup Parameter
-#'
-#' Validates that n_warmup is a non-negative integer
-#'
-#' @param n_warmup The number of warmup iterations
-#'
-#' @noRd
-#' @srrstats {G1.4} uses `Roxygen2` documentation
-#' @srrstats {G1.4a} internal function specified with `@noRd`
-#' @srrstats {G5.2a} every error statement is unique
-#'
-#' @return NULL (invisibly) if validation passes, otherwise stops with error
-validate_n_warmup <- function(n_warmup) {
-  if (!is.numeric(n_warmup)) {
-    cli::cli_abort("{.arg n_warmup} must be numeric, got {.cls {class(n_warmup)}}")
-  }
-
-  if (length(n_warmup) != 1) {
-    cli::cli_abort("{.arg n_warmup} must be a single value, got length {length(n_warmup)}")
-  }
-
-  if (!is.finite(n_warmup)) {
-    cli::cli_abort("{.arg n_warmup} must be a finite number")
-  }
-
[…]
