+++ R/validation-utils.R
@@ -102,6 +102,7 @@ validate_class_inherits <- function(obj, class_names, require_all = TRUE) {
 #' @noRd
 #' @srrstats {G1.4} uses `Roxygen2` documentation
 #' @srrstats {G1.4a} internal function specified with `@noRd`
+#' @srrstats {G2.6} one dimensional values are pre-processed regardless of class
 #' @srrstats {G5.2a} every error statement is unique
 #' @srrstats {G5.8, G5.8a} checks for zero length
 #' @srrstats {BS2.3, BS2.4, BS2.5} checks priors for length, appropriate
@@ -389,6 +390,64 @@ convert_ts_to_tsibble <- function(ts_obj) {
   return(temp_tsbl)
 }
 
+#' Check for list columns in data
+#'
+#' @param data A data.frame or data.frame-like object
+#' @param relevant_cols Character vector of column names that will be used
+#'
+#' @returns NULL (invisibly) if no list columns found, otherwise throws error
+#' @keywords internal
+#' @noRd
+#'
+#' @srrstats {G1.4} uses `Roxygen2` documentation
+#' @srrstats {G1.4a} internal function specified with `@noRd`
+#' @srrstats {G2.12} Check for list columns in data.frame-like objects and
+#'   provide informative error message
+#' @srrstats {G5.2a} every error statement is unique
+check_list_columns <- function(data, relevant_cols) {
+
+  # For time series objects, this check will happen after conversion to data.frame
+  # So we can safely assume data is data.frame-like here
+  if (!is.data.frame(data)) {
+    # Try to coerce to data.frame for checking
+    data <- tryCatch(
+      as.data.frame(data),
+      error = function(e) {
+        # If we can't convert to check, skip the check
+        # (will be handled by other validation)
+        return(NULL)
+      }
+    )
+    if (is.null(data)) return(invisible(NULL))
+  }
+
+  # Get columns that exist in the data
+  existing_cols <- intersect(relevant_cols, names(data))
+
+  if (length(existing_cols) == 0) {
+    return(invisible(NULL))
+  }
+
+  # Check for list columns among relevant columns
+  # Exclude data.frames (they are technically lists but have different semantics)
+  list_cols <- vapply(existing_cols, function(col) {
+    is.list(data[[col]]) && !is.data.frame(data[[col]])
+  }, logical(1))
+
+  if (any(list_cols)) {
+    list_col_names <- names(list_cols)[list_cols]
+
+    cli::cli_abort(c(
+      "List columns detected: {.val {list_col_names}}",
+      "x" = "List columns are not supported for time series analysis",
+      "i" = "Please convert these columns to appropriate vector types (numeric, character, etc.) or remove them from your data",
+      "i" = "You can use {.code unlist()} or extract specific elements if the list contains single values"
+    ))
+  }
+
+  invisible(NULL)
+}
+
 #' Helper: Create and validate tsibble from various data formats
 #'
 #' @param data input dataset - can be data frame, tibble, data.table, or time series object
@@ -399,6 +458,7 @@ convert_ts_to_tsibble <- function(ts_obj) {
 #' @srrstats {G1.4} uses `Roxygen2` documentation
 #' @srrstats {G1.4a} internal function specified with `@noRd`
 #' @srrstats {G5.2a} every error statement is unique
[…]
