### 2026-02-04 e2a20f8
remove already covered srr standards where have placed in other locations
