+++ R/srr-stats-standards.R
@@ -35,10 +35,7 @@ NULL
 #' @srrstatsNA {G1.6} This software makes no performance claims.
 #' @srrstatsNA {G2.4c} no conversion to character object
 #' @srrstatsNA {G2.4d, G2.4e, G2.5} no use of factor type objects
-#' @srrstatsNA {G2.6} vector input is confined to columns names, so this
-#'   pre-processing appears unnecessary
 #' @srrstatsNA {G2.9} no information is lost from type conversion to tsibble
-#' @srrstatsNA {G2.11, G2.12} data.frame-like objects are not used
 #' @srrstatsNA {G2.14b, G2.14c} These models do not currently accommodate data
 #'   with missing entries. So missing data cannot be ignored or replaced.
 #' @srrstatsNA {G3.1, G3.1a} The software does not rely on covariance calculations
@@ -50,42 +47,24 @@ NULL
 #'   instead uses functions from the `rstan` package.
 #' @srrstatsNA {G5.4b} This is not a new implementation of an existing method
 #' @srrstatsNA {G5.4c} No stored values are needed for correctness testing
-#' @srrstatsNA {G5.10,G5.11,G5.11a,G5.12} I have not implemented any extended
-#'   tests
 #' @srrstatsNA {BS1.0} The term 'hyperparameter' is not used
-#' @srrstatsNA {BS1.3a, BS2.8} Since all the sampling functions use `sampling`
-#'   from `rstan`, the user can specify initial values
 #' @srrstatsNA {BS1.3b} This software uses Stan that implements only one type
 #'   of algorithm (HMC/NUTS)
 #' @srrstatsNA {BS1.5} This software provides a helper function to diagnose
 #'   convergence, reporting the maximum R-hat and minimum n_eff. It does not
 #'   explicitly choose for the user if the model converged, and so comparison
-#'   of alternative dianostics is not necessary
+#'   of alternative diagnostics is not necessary
 #' @srrstatsNA {BS2.1, BS2.1a} only one tabular input, so numbers of rows don't
 #'   need to be compared between multiple inputs
-#' @srrstatsNA {BS2.6} Stan takes care of this directly
-#' @srrstatsNA {BS2.9} This software uses `rstan`'s `sampling`, which ensures
-#'   different seeds are used (from rstan's sampling() documentation: "Even if
-#'   multiple chains are used, only one seed is needed, with other chains having
-#'   seeds derived from that of the first chain to avoid dependent samples")
-#' @srrstatsNA {BS2.10} This software ensures that the same seed is not passed
-#'   to multiple computational chains
-#' @srrstatsNA {BS2.13, BS2.14} This software uses `rstan`'s `sampling`, and
-#'   while it seems possible to suppress all messages/warnings/progress bars, it
-#'   is unclear how to suppress just progress bars or just warnings.
-#' @srrstatsNA {BS3.1, BS3.2} These are not predictive models where there may be
-#'   risk of collinear response and predictor variables
+#' @srrstatsNA {BS3.1, BS3.2} These are not predictive models where there may
+#'   be risk of collinear response and predictor variables
 #' @srrstatsNA {BS4.0, BS4.1} This software does not implement internal
 #'  sampling algorithms
 #' @srrstatsNA {BS4.2, BS7.2} The only way to validate the posterior estimates
 #'  would be to compare against an analytical solution, which we don't have.
-#' @srrstatsNA {BS4.4} This software uses `rstan`'s `sampling`, and there
-#'   does not appear to yet be a mechanism of stopping the chain upon convergence
-#' @srrstatsNA {BS4.6, BS4.7, BS5.3, BS5.4} No separate convergence
-#'   checker is implemented. Stan software returns some convergence warnings
[…]
