### 2026-02-07 #738 @emilyriederer
Hello! I'm the rotating EiC for this quarter and doing a round of housekeeping on our issues. I'm closing this issue as the discussion will continue in https://github.com/ropensci/software-review/issues/749
