### 2026-03-03 eaae35f
add standards that stan qualifies
