### 2026-03-03 595228a
update time series related standards and add validation checks for mcmc params
