### 2026-03-04 44d2ac9
removed some standards previously listed as NA that were covered by stan
