### 2026-03-04 6bd7530
instead of creating fixtures as a separate step, generate them within gated environments that skip by default
