### 2026-03-04 8cca708
add small cached fitted models for some of the tests as the longer fitted models will not run/be available to them
