+++ tests/testthat/test-extended-regression.R
@@ -1,0 +1,219 @@
+# Regression tests for fitted models. These tests are slow because they
+# depend on model fitting (handled by helper-extended-fixtures.R).
+#
+# To run:
+#   EPISTRAINDYNAMICS_EXTENDED_TESTS=true Rscript -e 'devtools::test()'
+#
+#' @srrstats {G5.10} Extended tests included and switched on by flag
+#' @srrstats {G5.11, G5.11a} All data is package data (sarscov2, influenza),
+#'   fully reproducible without downloading external assets.
+#' @srrstats {G5.12} Requirements described in tests/README.md
+#'
+# ===========================================================================
+# Rt estimation
+# ===========================================================================
+
+test_that("Rt returns expected structure from random walk single-pathogen model", {
+  skip_extended()
+
+  gi_simple <- function(x) ifelse(x == 0, 0, 4 * x * exp(-2 * x))
+
+  result <- Rt(extended_fixtures$fit_rw_single, tau_max = 7, gi_dist = gi_simple)
+
+  # Structure checks
+  expect_true(!is.null(result$measure))
+  expect_true(nrow(result$measure) > 0)
+})
+
+test_that("Rt point estimates are finite and non-negative", {
+  skip_extended()
+
+  gi_simple <- function(x) ifelse(x == 0, 0, 4 * x * exp(-2 * x))
+
+  result <- Rt(extended_fixtures$fit_rw_single, tau_max = 7, gi_dist = gi_simple)
+
+  # All point estimates should be finite
+  expect_true(all(is.finite(result$measure$median)))
+
+  # Rt should be non-negative
+  expect_true(all(result$measure$median >= 0))
+})
+
+test_that("Rt credible intervals are ordered correctly", {
+  skip_extended()
+
+  gi_simple <- function(x) ifelse(x == 0, 0, 4 * x * exp(-2 * x))
+
+  result <- Rt(extended_fixtures$fit_rw_single, tau_max = 7, gi_dist = gi_simple)
+
+  # Lower bound <= median <= upper bound
+  expect_true(all(result$measure$lower <= result$measure$median))
+  expect_true(all(result$measure$median <= result$measure$upper))
+})
+
+# ===========================================================================
+# Incidence
+# ===========================================================================
+
+test_that("Incidence returns expected structure from random walk multi-pathogen model", {
+  skip_extended()
+
+  result <- incidence(extended_fixtures$fit_rw_multi, dow = FALSE)
+
+  expect_true(!is.null(result$measure))
+  expect_true(nrow(result$measure) > 0)
+})
+
+test_that("Incidence estimates are finite and non-negative", {
+  skip_extended()
+
+  result <- incidence(extended_fixtures$fit_rw_multi, dow = FALSE)
+
+  expect_true(all(is.finite(result$measure$median)))
+  expect_true(all(result$measure$median >= 0))
+})
+
+test_that("Incidence credible intervals are ordered correctly", {
+  skip_extended()
+
+  result <- incidence(extended_fixtures$fit_rw_multi, dow = FALSE)
+
+  expect_true(all(result$measure$lower <= result$measure$median))
+  expect_true(all(result$measure$median <= result$measure$upper))
+})
+
+# ===========================================================================
+# Growth rate
+# ===========================================================================
+
+test_that("Growth rate returns expected structure from random walk single-pathogen model", {
+  skip_extended()
+
+  result <- growth_rate(extended_fixtures$fit_rw_single)
+
+  expect_true(!is.null(result$measure))
+  expect_true(nrow(result$measure) > 0)
+})
+
+test_that("Growth rate estimates are finite", {
+  skip_extended()
+
+  result <- growth_rate(extended_fixtures$fit_rw_single)
+
+  # Growth rates can be negative, but should be finite
[…]
