### 2026-03-04 8812950
add tests for list-columns
