+++ tests/testthat/test-list_columns.R
@@ -1,0 +1,205 @@
+
+# Test helper function: check_list_columns
+test_that("check_list_columns detects list columns in data.frame", {
+  # Create data with list column
+  df <- data.frame(
+    time = 1:5,
+    cases = c(10, 20, 30, 40, 50),
+    metadata = I(list(
+      list(a = 1), list(a = 2), list(a = 3), list(a = 4), list(a = 5)
+    ))
+  )
+
+  expect_error(
+    check_list_columns(df, c("time", "cases", "metadata")),
+    "List columns detected"
+  )
+
+  expect_error(
+    check_list_columns(df, c("time", "cases", "metadata")),
+    "List columns are not supported for time series analysis"
+  )
+})
+
+test_that("check_list_columns passes when no list columns present", {
+  df <- data.frame(
+    time = 1:5,
+    cases = c(10, 20, 30, 40, 50),
+    name = c("a", "b", "c", "d", "e")
+  )
+
+  expect_silent(
+    check_list_columns(df, c("time", "cases", "name"))
+  )
+})
+
+test_that("check_list_columns ignores non-relevant list columns", {
+  df <- data.frame(
+    time = 1:5,
+    cases = c(10, 20, 30, 40, 50),
+    metadata = I(list(
+      list(a = 1), list(a = 2), list(a = 3), list(a = 4), list(a = 5)
+    ))
+  )
+
+  # Should pass because metadata is not in relevant_cols
+  expect_silent(
+    check_list_columns(df, c("time", "cases"))
+  )
+})
+
+test_that("check_list_columns handles multiple list columns", {
+  df <- data.frame(
+    time = 1:5,
+    cases = c(10, 20, 30, 40, 50),
+    metadata1 = I(list(
+      list(a = 1), list(a = 2), list(a = 3), list(a = 4), list(a = 5)
+    )),
+    metadata2 = I(list(
+      list(b = 1), list(b = 2), list(b = 3), list(b = 4), list(b = 5)
+    ))
+  )
+
+  expect_error(
+    check_list_columns(df, c("time", "cases", "metadata1", "metadata2")),
+    "List columns detected"
+  )
+})
+
+test_that("check_list_columns handles tibbles with list columns", {
+  skip_if_not_installed("tibble")
+  library(tibble)
+
+  df <- tibble(
+    time = 1:5,
+    cases = c(10, 20, 30, 40, 50),
+    metadata = list(
+      list(a = 1), list(a = 2), list(a = 3), list(a = 4), list(a = 5)
+    )
+  )
+
+  expect_error(
+    check_list_columns(df, c("time", "cases", "metadata")),
+    "List columns detected"
+  )
+})
+
+# Test single() function with list columns
+test_that("single() rejects data.frame with list columns", {
+  df <- data.frame(
+    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 5),
+    cases = c(10, 20, 30, 40, 50),
+    metadata = I(list(
+      list(a = 1), list(a = 2), list(a = 3), list(a = 4), list(a = 5)
+    ))
+  )
+
+  expect_error(
+    single(data = df, case_timeseries = "metadata", time = "date"),
+    "List columns detected"
+  )
+})
+
+# Test multiple() function with list columns
+test_that("multiple() rejects data.frame with list columns in
+          component_pathogen_timeseries", {
+  df <- data.frame(
+    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 5),
+    cases = c(100, 200, 300, 400, 500),
+    alpha = c(10, 20, 30, 40, 50),
+    delta = I(list(
[…]
