### 2026-03-04 7ae843b
new helper functions for method tests
