### 2026-03-04 ec66ef5
tidy formatting and add tests for units package support
