+++ tests/testthat/test-units-pkg-support.R
@@ -1,35 +1,35 @@
-# Tests for units package support
+# Basic Units Support Tests
 
-test_that("single() accepts units objects in case timeseries", {
+test_that("single() accepts and processes units objects ", {
   skip_if_not_installed("units")
+  library(units)
 
   df <- data.frame(
-    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 10),
-    cases = units::set_units(1:10, "1")  # Dimensionless count
+    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 5)
   )
+  df$cases <- as_units(c(10, 20, 30, 40, 50), "count")
 
-  result <- single(
-    data = df,
-    case_timeseries = "cases",
-    time = "date"
-  )
+  # Should work without error
+  result <- single(data = df, case_timeseries = "cases", time = "date")
 
   expect_s3_class(result, "EpiStrainDynamics.pathogen_structure")
   expect_equal(result$pathogen_structure, "single")
-  # Units should be stripped, result should be numeric
-  expect_true(is.numeric(result$data$case_timeseries))
-  expect_equal(length(result$data$case_timeseries), 10)
+
+  # Units should be converted to numeric
+  expect_type(result$data$case_timeseries, "double")
+  expect_equal(as.numeric(result$data$case_timeseries), c(10, 20, 30, 40, 50))
 })
 
-test_that("multiple() accepts units objects in component pathogens", {
+test_that("multiple() handles units objects in case_timeseries ", {
   skip_if_not_installed("units")
+  library(units)
 
   df <- data.frame(
-    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 10),
-    cases = units::set_units(1:10, "1"),
-    alpha = units::set_units(11:20, "1"),
-    delta = units::set_units(21:30, "1")
+    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 5),
+    alpha = c(10, 20, 30, 40, 50),
+    delta = c(5, 10, 15, 20, 25)
   )
+  df$cases <- as_units(c(100, 200, 300, 400, 500), "count")
 
   result <- multiple(
     data = df,
@@ -39,51 +39,76 @@ test_that("multiple() accepts units objects in component pathogens", {
   )
 
   expect_s3_class(result, "EpiStrainDynamics.pathogen_structure")
-  expect_equal(result$pathogen_structure, "multiple")
-  # Units should be stripped
-  expect_true(is.numeric(result$data$case_timeseries))
-  expect_true(is.numeric(result$data$component_pathogens))
+  expect_type(result$data$case_timeseries, "double")
+  expect_equal(as.numeric(result$data$case_timeseries), c(100, 200, 300, 400, 500))
 })
 
-test_that("subtyped() accepts units objects in all pathogen timeseries", {
+test_that("multiple() handles units objects in component pathogens ", {
   skip_if_not_installed("units")
+  library(units)
 
   df <- data.frame(
-    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 10),
-    ili = units::set_units(seq(100, 190, by = 10), "1"),
-    inf_A = units::set_units(seq(30, 75, by = 5), "1"),
-    inf_H3N2 = units::set_units(seq(10, 28, by = 2), "1"),
-    inf_H1N1 = units::set_units(seq(8, 26, by = 2), "1"),
-    inf_B = units::set_units(seq(20, 47, by = 3), "1"),
-    other = units::set_units(seq(5, 14, by = 1), "1")
+    date = seq.Date(as.Date("2020-01-01"), by = "day", length.out = 5),
+    cases = c(100, 200, 300, 400, 500)
   )
[…]
