### 2026-03-04 6b91d07
add tests for mcmc param validation
