### 2026-03-04 6b4111f
add tests for verbosity of fit model
