+++ tests/testthat/test-verbosity.R
@@ -1,0 +1,300 @@
+# Tests for verbosity and error handling
+# These tests verify that fit_model properly handles verbose output,
+# message/warning suppression, and error catching per Bayesian standards.
+#
+# Note: Core fit_model functionality is tested in test-fit-model.R
+
+test_that("fit_model has verbose parameter defaulting to TRUE", {
+  # Check that the function has verbose argument
+  fit_model_args <- names(formals(fit_model))
+  expect_true("verbose" %in% fit_model_args)
+
+  # Check that default is TRUE
+  expect_true(formals(fit_model)$verbose)
+})
+
+test_that("fit_model with verbose=TRUE shows progress", {
+  skip_on_cran()
+  skip_if_not_installed("rstan")
+  check_package_data()
+
+  sarscov2_subset <- sarscov2[1:40, ]
+
+  mod <- construct_model(
+    method = random_walk(),
+    pathogen_structure = single(
+      data = sarscov2_subset,
+      case_timeseries = "cases",
+      time = "date"
+    )
+  )
+
+  # With verbose=TRUE, function should complete successfully
+  suppressWarnings(
+    fit <- fit_model(mod, n_iter = 100, n_chain = 1, verbose = TRUE)
+  )
+
+  expect_s3_class(fit, "EpiStrainDynamics.fit")
+})
+
+# Test suppress messages/progress but retain warnings/errors
+
+test_that("fit_model has separate control for messages and warnings", {
+  # verbose parameter should control messages/progress
+  # warnings should still appear even when verbose=FALSE
+  fit_model_args <- names(formals(fit_model))
+  expect_true("verbose" %in% fit_model_args)
+  expect_true("suppress_warnings" %in% fit_model_args)
+})
+
+test_that("fit_model with verbose=FALSE suppresses progress", {
+  skip_on_cran()
+  skip_if_not_installed("rstan")
+  check_package_data()
+
+  sarscov2_subset <- sarscov2[1:40, ]
+
+  mod <- construct_model(
+    method = random_walk(),
+    pathogen_structure = single(
+      data = sarscov2_subset,
+      case_timeseries = "cases",
+      time = "date"
+    )
+  )
+
+  # Capture output with verbose=FALSE
+  output <- capture.output({
+    suppressWarnings(
+      fit <- fit_model(mod, n_iter = 100, n_chain = 1, verbose = FALSE)
+    )
+  }, type = "message")
+
+  # Should complete successfully
+  expect_s3_class(fit, "EpiStrainDynamics.fit")
+
+  # Progress messages should be minimal
+  expect_lt(length(output), 5)
+})
+
+# Test suppress warnings when appropriate
+
+test_that("fit_model has suppress_warnings parameter", {
+  fit_model_args <- names(formals(fit_model))
+  expect_true("suppress_warnings" %in% fit_model_args)
+
+  # Check that default is FALSE
+  expect_false(formals(fit_model)$suppress_warnings)
+})
+
+test_that("fit_model with suppress_warnings=TRUE suppresses warnings", {
+  skip_on_cran()
+  skip_if_not_installed("rstan")
+  check_package_data()
+
+  sarscov2_subset <- sarscov2[1:40, ]
+
+  mod <- construct_model(
+    method = random_walk(),
+    pathogen_structure = single(
+      data = sarscov2_subset,
+      case_timeseries = "cases",
+      time = "date"
+    )
+  )
+
+  # With suppress_warnings=TRUE, warnings should be suppressed
+  expect_no_error({
+    fit <- fit_model(
+      mod,
+      n_iter = 100,
+      n_chain = 1,
[…]
