### 2026-03-05 a5e854c
conflicting cli message and remove redundant tests
