+++ articles/algorithmic-scaling.md
@@ -290,10 +290,10 @@ knitr::kable(scaling_results, digits = 3,
 
 | method      | pathogen_structure | scaling_exponent | r_squared |
 |:------------|:-------------------|-----------------:|----------:|
-| p_spline    | multiple           |            1.014 |     0.997 |
-| p_spline    | single             |            0.855 |     0.991 |
-| random_walk | multiple           |            0.964 |     0.999 |
-| random_walk | single             |            0.890 |     0.985 |
+| p_spline    | multiple           |            1.010 |     0.996 |
+| p_spline    | single             |            0.853 |     0.991 |
+| random_walk | multiple           |            0.967 |     0.999 |
+| random_walk | single             |            0.887 |     0.985 |
 
 Scaling exponents: time complexity approximately O(n^exponent)
 
@@ -806,30 +806,39 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
-#> [1] dplyr_1.1.4                  ggplot2_4.0.1               
+#> [1] dplyr_1.2.0                  ggplot2_4.0.2               
 #> [3] EpiStrainDynamics_0.0.1.0000
 #> 
 #> loaded via a namespace (and not attached):
-#>  [1] viridis_0.6.5         sass_0.4.10           generics_0.1.4       
-#>  [4] anytime_0.3.12        digest_0.6.39         magrittr_2.0.4       
-#>  [7] timechange_0.3.0      evaluate_1.0.5        grid_4.5.2           
-#> [10] RColorBrewer_1.1-3    fastmap_1.2.0         jsonlite_2.0.0       
-#> [13] pkgbuild_1.4.8        gridExtra_2.3         purrr_1.2.0          
-#> [16] viridisLite_0.4.2     QuickJSR_1.8.1        scales_1.4.0         
-#> [19] codetools_0.2-20      textshaping_1.0.4     jquerylib_0.1.4      
-#> [22] cli_3.6.5             rlang_1.1.6           ellipsis_0.3.2       
-#> [25] splines_4.5.2         withr_3.0.2           cachem_1.1.0         
-#> [28] yaml_2.3.12           StanHeaders_2.32.10   tools_4.5.2          
-#> [31] rstan_2.32.7          inline_0.3.21         parallel_4.5.2       
-#> [34] rstantools_2.5.0      tsibble_1.1.6         vctrs_0.6.5          
-#> [37] R6_2.6.1              lubridate_1.9.4       matrixStats_1.5.0    
-#> [40] stats4_4.5.2          lifecycle_1.0.4       fs_1.6.6             
-#> [43] ragg_1.5.0            pkgconfig_2.0.3       desc_1.4.3           
-#> [46] pkgdown_2.2.0         RcppParallel_5.1.11-1 pillar_1.11.1        
-#> [49] bslib_0.9.0           gtable_0.3.6          loo_2.9.0            
-#> [52] glue_1.8.0            Rcpp_1.1.0            systemfonts_1.3.1    
-#> [55] xfun_0.55             tibble_3.3.0          tidyselect_1.2.1     
-#> [58] knitr_1.51            farver_2.1.2          bayesplot_1.15.0     
-#> [61] htmltools_0.5.9       labeling_0.4.3        rmarkdown_2.30       
-#> [64] compiler_4.5.2        S7_0.2.1
+#>  [1] tidyselect_1.2.1      viridisLite_0.4.3     timeDate_4052.112    
+#>  [4] farver_2.1.2          viridis_0.6.5         loo_2.9.0            
[…]
