+++ articles/Using-EpiStrainDynamics.md
@@ -56,16 +56,22 @@ There are three main types of pathogen structure available to model:
 and
 [`subtyped()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/subtyped.md).
 These functions require the name of the dataset itself and the column
-names for different data elements.
+names for different data elements. These can also ingest timeseries
+class objects, which would then make the `time` column specification
+optional. The models are not currently equipped to handle missing values
+or irregular date sequences. Timeseries checks are run internally on
+data frame or any time of timeseries class input that is handled by the
+[`tsbox`
+package](https://docs.ropensci.org/tsbox/reference/tsbox-package.html).
 
 The
 [`single()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/single.md)
 pathogen structure is the simplest and models a single pathogen
 timeseries. The name of the dataframe is passed to argument `data`, the
 name of the column with total case data is passed to `case_timeseries`,
-and the name of the column of time data is passed to `time`. It can be
-specified as follows, illustrated using data provided with the package
-`sarscov2`:
+and the name of the column of time data is passed to `time` (optional if
+input dataset is a timeseries class). It can be specified as follows,
+illustrated using data provided with the package `sarscov2`:
 
     pathogen_structure = single(
       data = sarscov2,                    # dataframe
@@ -76,9 +82,10 @@ specified as follows, illustrated using data provided with the package
 The
 [`multiple()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/multiple.md)
 pathogen structure allows modelling of different component pathogens. In
-addition to specifying `data`, `case_timeseries`, and `time`, these
-additional pathogens are specified as a vector of column names with the
-argument `component_pathogen_timeseries`. Example pathogen structure
+addition to specifying `data`, `case_timeseries`, and `time` (optional
+if input dataset is a timeseries class), these additional pathogens are
+specified as a vector of column names with the argument
+`component_pathogen_timeseries`. Example pathogen structure
 specification for multiple pathogens model using the `sarscov2` dataset:
 
     pathogen_structure = multiple(
@@ -188,12 +195,15 @@ mod <- construct_model(
 
 ### Output
 
-The constructed model object is a named list containing input data,
-accessed with `$data`, parameter values (such as smoothing structure,
-observation noise, and penalised spline parameters, if appropriate),
-accessed with `model_params`, names of provided pathogens, accessed with
-`pathogen_names`, and record of whether day of week effect has been
-selected, accessed with `dow_effect`.
+The constructed model object is a named list containing the formatted
+data to be modelled, accessed with `$data`, a timeseries object of class
+`tsibble` that has been validated for gaps and regularity, accessed with
+`$validated_tsbl`, parameter values structured for the stan interface
+(such as smoothing structure, observation noise, and penalised spline
+parameters, if appropriate), accessed with `model_params`, names of
+provided pathogens, accessed with `pathogen_names`, and record of
+whether day of week effect has been selected, accessed with
+`dow_effect`.
 
 ## Step 2: Fit model
 
@@ -227,8 +237,8 @@ diagnose_model(fit)
[…]
