+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/algorithmic-scaling: algorithmic-scaling.html
   articles/parameter-recovery: parameter-recovery.html
   Using-EpiStrainDynamics: Using-EpiStrainDynamics.html
-last_built: 2026-03-06T02:03Z
+last_built: 2026-03-06T15:55Z
 urls:
   reference: https://acefa-hubs.github.io/EpiStrainDynamics/reference
   article: https://acefa-hubs.github.io/EpiStrainDynamics/articles
