### 2026-03-08 c944f2d
add explicit message to skip extended tests
