+++ .github/workflows/codecov.yaml
@@ -4,29 +4,23 @@ on:
   push:
     branches: main
   pull_request:
-
 name: codecov.yaml
-
 permissions: read-all
-
 jobs:
   test-coverage:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
+      EPISTRAINDYNAMICS_EXTENDED_TESTS: "false"
     steps:
       - uses: actions/checkout@v4
-
       - uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
-
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::covr, any::xml2
           needs: coverage
-
       - name: Test coverage
         run: |
           cov <- covr::package_coverage(
@@ -37,7 +31,6 @@ jobs:
           )
           covr::to_cobertura(cov)
         shell: Rscript {0}
-
       - uses: codecov/codecov-action@v4
         with:
           # Fail if error if not on PR, or if on PR and token is given
@@ -46,18 +39,15 @@ jobs:
           plugin: noop
           disable_search: true
           token: ${{ secrets.CODECOV_TOKEN }}
-
       - name: Show testthat output
         if: always()
         run: |
           ## --------------------------------------------------------------------
           find '${{ runner.temp }}/package' -name 'testthat.Rout*' -exec cat '{}' \; || true
         shell: bash
-
       - name: Upload test results
         if: failure()
         uses: actions/upload-artifact@v4
         with:
           name: coverage-test-failures
           path: ${{ runner.temp }}/package
-
