+++ articles/algorithmic-scaling.md
@@ -290,10 +290,10 @@ knitr::kable(scaling_results, digits = 3,
 
 | method      | pathogen_structure | scaling_exponent | r_squared |
 |:------------|:-------------------|-----------------:|----------:|
-| p_spline    | multiple           |            1.009 |     0.995 |
-| p_spline    | single             |            0.845 |     0.993 |
-| random_walk | multiple           |            0.963 |     0.999 |
-| random_walk | single             |            0.869 |     0.980 |
+| p_spline    | multiple           |            1.009 |     0.996 |
+| p_spline    | single             |            0.851 |     0.991 |
+| random_walk | multiple           |            0.967 |     0.999 |
+| random_walk | single             |            0.885 |     0.981 |
 
 Scaling exponents: time complexity approximately O(n^exponent)
 
@@ -824,7 +824,7 @@ sessionInfo()
 #> [34] xts_0.14.2            future_1.69.0         inline_0.3.21        
 #> [37] globals_0.19.0        scales_1.4.0          MASS_7.3-65          
 #> [40] anytime_0.3.12        cli_3.6.5             rmarkdown_2.30       
-#> [43] ragg_1.5.0            generics_0.1.4        otel_0.2.0           
+#> [43] ragg_1.5.1            generics_0.1.4        otel_0.2.0           
 #> [46] RcppParallel_5.1.11-2 future.apply_1.20.2   cachem_1.1.0         
 #> [49] rstan_2.32.7          splines_4.5.2         bayesplot_1.15.0     
 #> [52] parallel_4.5.2        matrixStats_1.5.0     vctrs_0.7.1          
@@ -836,9 +836,9 @@ sessionInfo()
 #> [70] gtable_0.3.6          QuickJSR_1.9.0        tibble_3.3.1         
 #> [73] pillar_1.11.1         furrr_0.3.1           htmltools_0.5.9      
 #> [76] ipred_0.9-15          lava_1.8.2            R6_2.6.1             
-#> [79] textshaping_1.0.4     evaluate_1.0.5        lattice_0.22-7       
+#> [79] textshaping_1.0.5     evaluate_1.0.5        lattice_0.22-7       
 #> [82] tsibble_1.2.0         bslib_0.10.0          rstantools_2.6.0     
 #> [85] class_7.3-23          Rcpp_1.1.1            gridExtra_2.3        
-#> [88] prodlim_2025.04.28    xfun_0.56             fs_1.6.6             
+#> [88] prodlim_2025.04.28    xfun_0.56             fs_1.6.7             
 #> [91] zoo_1.8-15            pkgconfig_2.0.3
 ```
