### 2026-03-09 ed0aaa4
try again to set extended tests to false
