+++ tests/testthat/setup-fitted-models.R
@@ -1,166 +1,41 @@
-# Files that use these: test-fit-model.R, test-diagnose-model.R,
-# test-metrics.R, test-plot.R
+# tests/testthat/helper-fixtures.R
+# Downloads pre-fitted model fixtures from GitHub Release via piggyback.
+# Fitting locally is skipped entirely on CI.
 
-# Only create fitted models if package data is available
 if (exists("sarscov2") && exists("influenza")) {
 
-  message("Setting up cached fitted models for test session...")
+  fixture_dir <- file.path(tempdir(), "testfixtures")
+  dir.create(fixture_dir, showWarnings = FALSE)
 
-  # Suppress Stan output during setup
-  options(browser = function(...) {})
+  message("Downloading pre-fitted model fixtures via piggyback...")
 
-  # Small subsets for speed
-  sarscov2_subset <- sarscov2[1:50, ]
-  influenza_subset <- influenza[1:40, ]
-
-  # Create and cache fitted models in the global test environment
-  # These will be available to ALL test files
-
-  fit_rw_single <<- suppressWarnings(
-    fit_model(
-      construct_model(
-        method = random_walk(),
-        pathogen_structure = single(
-          data = sarscov2_subset,
-          case_timeseries = 'cases',
-          time = 'date'
-        )
-      ),
-      n_iter = 500,
-      n_chain = 1,
-      verbose = FALSE,
-      seed = 123
-    )
+  piggyback::pb_download(
+    repo    = "your-org/EpiStrainDynamics",
+    tag     = "test-fixtures-v1",
+    dest    = fixture_dir,
+    overwrite = TRUE
   )
 
-  fit_ps_single <<- suppressWarnings(
-    fit_model(
-      construct_model(
-        method = p_spline(),
-        pathogen_structure = single(
-          data = sarscov2_subset,
-          case_timeseries = 'cases',
-          time = 'date'
-        )
-      ),
-      n_iter = 500,
-      n_chain = 1,
-      verbose = FALSE,
-      seed = 123
-    )
-  )
+  # Load into global test environment
+  fit_rw_single     <<- readRDS(file.path(fixture_dir, "fit_rw_single.rds"))
+  fit_ps_single     <<- readRDS(file.path(fixture_dir, "fit_ps_single.rds"))
+  fit_rw_multi      <<- readRDS(file.path(fixture_dir, "fit_rw_multi.rds"))
+  fit_ps_multi      <<- readRDS(file.path(fixture_dir, "fit_ps_multi.rds"))
+  fit_rw_subtyped   <<- readRDS(file.path(fixture_dir, "fit_rw_subtyped.rds"))
+  fit_ps_subtyped   <<- readRDS(file.path(fixture_dir, "fit_ps_subtyped.rds"))
+  fit_rw_single_dow <<- readRDS(file.path(fixture_dir, "fit_rw_single_dow.rds"))
+  inc_single        <<- readRDS(file.path(fixture_dir, "inc_single.rds"))
+  inc_multi         <<- readRDS(file.path(fixture_dir, "inc_multi.rds"))
+  gr_single         <<- readRDS(file.path(fixture_dir, "gr_single.rds"))
+  gr_multi          <<- readRDS(file.path(fixture_dir, "gr_multi.rds"))
+  rt_single         <<- readRDS(file.path(fixture_dir, "rt_single.rds"))
+  rt_multi          <<- readRDS(file.path(fixture_dir, "rt_multi.rds"))
+  prop              <<- readRDS(file.path(fixture_dir, "prop.rds"))
 
-  fit_rw_multi <<- suppressWarnings(
-    fit_model(
-      construct_model(
-        method = random_walk(),
-        pathogen_structure = multiple(
[…]
