### 2026-03-17 9b8bf1b
bring timeout back down and remove the extended test flag now that they are downloaded from piggybank
