+++ inst/scripts/generate-test-fixtures.R
@@ -1,0 +1,248 @@
+# inst/scripts/generate-test-fixtures.R
+#
+# Generates and saves all test fixtures (regular and extended) to
+# inst/testfixtures/ and uploads them to the GitHub Release via piggyback.
+#
+# Run this script locally whenever fixtures need to be regenerated (e.g. after
+# model structure changes). Do NOT run on CI.
+#
+# Usage:
+#   source("inst/scripts/generate-test-fixtures.R")
+#
+# Requirements:
+#   - Run from the package root directory
+#   - Package must be loaded: devtools::load_all()
+#   - GitHub PAT with repo scope set in environment
+
+library(piggyback)
+
+# ------------------------------------------------------------------------------
+# Checks
+# ------------------------------------------------------------------------------
+
+if (!dir.exists("inst/testfixtures")) {
+  stop("inst/testfixtures does not exist - check your working directory")
+}
+
+if (!exists("sarscov2") || !exists("influenza")) {
+  stop("Package data (sarscov2, influenza) not available - run devtools::load_all() first")
+}
+
+# ------------------------------------------------------------------------------
+# Regular fixtures
+# (small data subsets, minimal MCMC settings for fast CI tests)
+# ------------------------------------------------------------------------------
+
+message("=== Generating regular fixtures ===")
+
+sarscov2_subset  <- sarscov2[1:50, ]
+influenza_subset <- influenza[1:40, ]
+
+message("  Fitting: Random walk, single pathogen...")
+fit_rw_single <- suppressWarnings(fit_model(
+  construct_model(
+    method = random_walk(),
+    pathogen_structure = single(
+      data = sarscov2_subset, case_timeseries = "cases", time = "date"
+    )
+  ),
+  n_iter = 500, n_chain = 1, verbose = FALSE, seed = 123
+))
+saveRDS(fit_rw_single, "inst/testfixtures/fit_rw_single.rds")
+
+message("  Fitting: P-spline, single pathogen...")
+fit_ps_single <- suppressWarnings(fit_model(
+  construct_model(
+    method = p_spline(),
+    pathogen_structure = single(
+      data = sarscov2_subset, case_timeseries = "cases", time = "date"
+    )
+  ),
+  n_iter = 500, n_chain = 1, verbose = FALSE, seed = 123
+))
+saveRDS(fit_ps_single, "inst/testfixtures/fit_ps_single.rds")
+
+message("  Fitting: Random walk, multiple pathogens...")
+fit_rw_multi <- suppressWarnings(fit_model(
+  construct_model(
+    method = random_walk(),
+    pathogen_structure = multiple(
+      data = sarscov2_subset, case_timeseries = "cases", time = "date",
+      component_pathogen_timeseries = c("alpha", "delta", "omicron", "other")
+    )
+  ),
+  n_iter = 500, n_chain = 1, verbose = FALSE, seed = 123
+))
+saveRDS(fit_rw_multi, "inst/testfixtures/fit_rw_multi.rds")
+
+message("  Fitting: P-spline, multiple pathogens...")
+fit_ps_multi <- suppressWarnings(fit_model(
+  construct_model(
+    method = p_spline(),
+    pathogen_structure = multiple(
+      data = sarscov2_subset, case_timeseries = "cases", time = "date",
+      component_pathogen_timeseries = c("alpha", "delta", "omicron", "other")
+    )
+  ),
[…]

[…]
