### 2026-03-17 c97cecd
ensure warnings are suppressed at appropriate time
