### 2026-03-17 a4cd3d2
revise readme given new extended test settings
