+++ tests/README.md
@@ -3,60 +3,62 @@
 ## Running tests
 
 Standard tests run with the usual `devtools::test()` or `R CMD check`. No
-special configuration is needed.
+special configuration is needed. All tests, including regression tests against
+pre-fitted models, run by default.
 
-## Extended tests
+## How fixtures work
 
-A subset of tests fit Stan models and are therefore slow (several hours on a
-typical machine). These tests are **skipped by default** and are gated behind
-an environment variable so that they can be opted into deliberately.
+Rather than fitting Stan models during the test suite (which is slow), all
+fitted model fixtures are pre-generated locally and stored in a GitHub Release.
+They are downloaded automatically at the start of each test session via
+[piggyback](https://docs.ropensci.org/piggyback/).
 
-### Enabling extended tests locally
+Two sets of fixtures are used:
 
-Set the environment variable **before** starting R (or export it in your shell):
+- **Regular fixtures** (`helper-fixtures.R`): small data subsets, minimal MCMC
+  settings (500 iterations, 1 chain). Used by `test-fit-model.R`,
+  `test-diagnose-model.R`, `test-metrics.R`, and `test-plot.R`.
+- **Extended fixtures** (`helper-extended-fixtures.R`): full package datasets,
+  thorough MCMC settings (2000 iterations, 4 chains). Used by
+  `test-extended-regression.R`.
 
-  ```bash
-export EPISTRAINDYNAMICS_EXTENDED_TESTS=true
-Rscript -e "devtools::test()"
-```
+Both helpers are sourced automatically by testthat before any tests run.
+Fixtures live in memory for the duration of the test session and are discarded
+afterwards.
 
-Or set it within an R session before calling `devtools::test()`:
+## Regenerating fixtures
 
+Fixtures only need to be regenerated when model structure changes in a way that
+affects the fitted object format. To regenerate:
 ```r
-Sys.setenv(EPISTRAINDYNAMICS_EXTENDED_TESTS = "true")
-devtools::test() # OR devtools::test(filter = "extended")
+devtools::load_all()
+source("inst/scripts/generate-test-fixtures.R")
 ```
 
-### How the extended tests work
+This fits all regular and extended fixtures locally, saves them to
+`inst/testfixtures/`, and uploads them to the GitHub Release. Bump the tag
+(e.g. `test-fixtures-v2`) and update the tag name in both helper files and the
+generate script when doing so.
 
-1. `tests/testthat/helper-extended-fixtures.R` is sourced automatically by
-testthat before any tests run. When the environment variable is set it fits
-all fixture models (random walk and p-spline, single and multiple pathogen)
-and stores the results in a shared in-memory environment called
-`extended_fixtures`.
+## Adding new tests
 
-2. `tests/testthat/test-extended-regression.R` contains the actual test cases.
-Each test calls `skip_extended()` at the top; if the environment variable is
-not set the entire file is skipped with a single diagnostic message.
+### Standard tests
+Add test cases to the appropriate `test-*.R` file. If the test needs a new
+fitted model, add it to `inst/scripts/generate-test-fixtures.R`, regenerate,
+and load it in `helper-fixtures.R`.
 
-3. The fixtures use package datasets so are fully reproducible. No external 
-data downloads are required.
+### Extended regression tests
+Add test cases to `test-extended-regression.R` (or a new `test-extended-*.R`
+file). If the tests need a new fitted model, add it to the extended fixtures
+section of `inst/scripts/generate-test-fixtures.R`, regenerate, and load it in
+`helper-extended-fixtures.R`.
 
-### Requirements and expectations
+## Requirements
 
 | Requirement | Detail |
-  |---|---|
[…]
