### 2026-03-17 538c4a8
revise tidyverse dependency
