### 2026-03-17 e68fcc6
revise extended fixtures
