### 2026-03-17 66c80eb
make iterations larger so that correct error is flagged
