### 2026-03-17 0c616b7
remove overly slow test
