+++ tests/testthat/helper-fixtures.R
@@ -18,9 +18,9 @@ dir.create(fixture_dir, showWarnings = FALSE)
 
 download_ok <- tryCatch({
   piggyback::pb_download(
-    repo      = "acefa-hubs/epistraindynamics",
-    tag       = "test-fixtures-v1",
-    dest      = fixture_dir,
+    repo = "acefa-hubs/epistraindynamics",
+    tag = "test-fixtures-v1",
+    dest = fixture_dir,
     overwrite = TRUE
   )
   TRUE
@@ -30,33 +30,40 @@ download_ok <- tryCatch({
 })
 
 if (!download_ok) {
-  skip("Fixture download from GitHub Release failed. Check network access and GitHub PAT.")
+  skip("Fixture download from GitHub Release failed. Check network access and
+       GitHub PAT.")
 }
 
 # Regular fixtures (small data subsets, minimal MCMC settings)
-fit_rw_single     <<- readRDS(file.path(fixture_dir, "fit_rw_single.rds"))
-fit_ps_single     <<- readRDS(file.path(fixture_dir, "fit_ps_single.rds"))
-fit_rw_multi      <<- readRDS(file.path(fixture_dir, "fit_rw_multi.rds"))
-fit_ps_multi      <<- readRDS(file.path(fixture_dir, "fit_ps_multi.rds"))
-fit_rw_subtyped   <<- readRDS(file.path(fixture_dir, "fit_rw_subtyped.rds"))
-fit_ps_subtyped   <<- readRDS(file.path(fixture_dir, "fit_ps_subtyped.rds"))
+fit_rw_single <<- readRDS(file.path(fixture_dir, "fit_rw_single.rds"))
+fit_ps_single <<- readRDS(file.path(fixture_dir, "fit_ps_single.rds"))
+fit_rw_multi <<- readRDS(file.path(fixture_dir, "fit_rw_multi.rds"))
+fit_ps_multi <<- readRDS(file.path(fixture_dir, "fit_ps_multi.rds"))
+fit_rw_subtyped <<- readRDS(file.path(fixture_dir, "fit_rw_subtyped.rds"))
+fit_ps_subtyped <<- readRDS(file.path(fixture_dir, "fit_ps_subtyped.rds"))
 fit_rw_single_dow <<- readRDS(file.path(fixture_dir, "fit_rw_single_dow.rds"))
-gi_simple         <<- readRDS(file.path(fixture_dir, "gi_simple.rds"))
-inc_single        <<- readRDS(file.path(fixture_dir, "inc_single.rds"))
-inc_multi         <<- readRDS(file.path(fixture_dir, "inc_multi.rds"))
-gr_single         <<- readRDS(file.path(fixture_dir, "gr_single.rds"))
-gr_multi          <<- readRDS(file.path(fixture_dir, "gr_multi.rds"))
-rt_single         <<- readRDS(file.path(fixture_dir, "rt_single.rds"))
-rt_multi          <<- readRDS(file.path(fixture_dir, "rt_multi.rds"))
-prop              <<- readRDS(file.path(fixture_dir, "prop.rds"))
+gi_simple <<- readRDS(file.path(fixture_dir, "gi_simple.rds"))
+inc_single <<- readRDS(file.path(fixture_dir, "inc_single.rds"))
+inc_multi <<- readRDS(file.path(fixture_dir, "inc_multi.rds"))
+gr_single <<- readRDS(file.path(fixture_dir, "gr_single.rds"))
+gr_multi <<- readRDS(file.path(fixture_dir, "gr_multi.rds"))
+rt_single <<- readRDS(file.path(fixture_dir, "rt_single.rds"))
[…]
