### 2026-03-17 f056450
ensure tests are skipped if fixture download fails
