### 2026-03-17 f490368
add check and dependency for piggyback
