### 2026-03-17 ea127bf
break up function with high cyclomatic complexity
