+++ R/validation-utils.R
@@ -161,74 +161,94 @@ validate_priors <- function(mean, sd) {
 #' @srrstats {G5.8, G5.8a} checks for zero length
 validate_smoothing_structure <- function(smoothing_obj, pathogen_names = NULL,
                                          pathogen_str) {
-  # Check if object has correct class
   if (!inherits(smoothing_obj, "EpiStrainDynamics.smoothing")) {
     stop("smoothing_params must be created using the smoothing_structure() function")
   }
 
-  if (pathogen_str == 'single' & smoothing_obj$smoothing_type != 'shared') {
-    smoothing_obj$smoothing_type <- 'shared'
+  smoothing_obj <- validate_smoothing_for_single(smoothing_obj, pathogen_str)
+  smoothing_obj <- validate_smoothing_independent(smoothing_obj, pathogen_names)
+  smoothing_obj <- validate_smoothing_shared(smoothing_obj)
+  smoothing_obj <- validate_smoothing_correlated(smoothing_obj)
+
+  return(smoothing_obj)
+}
+
+#' @noRd
+validate_smoothing_for_single <- function(smoothing_obj, pathogen_str) {
+  if (pathogen_str != "single") return(smoothing_obj)
+
+  if (smoothing_obj$smoothing_type != "shared") {
+    smoothing_obj$smoothing_type <- "shared"
     cli::cli_alert('smoothing_type can only be "shared" for "single" pathogen_type')
   }
-  if (pathogen_str == 'single' & is.null(smoothing_obj$tau_priors$mean)){
+  if (is.null(smoothing_obj$tau_priors$mean)) {
     smoothing_obj$tau_priors$mean <- 0.0
     smoothing_obj$tau_priors$sd <- 1.0
-    return(smoothing_obj)
   }
+  return(smoothing_obj)
+}
 
-  # For independent structure, validate and adjust dimensions
-  if (smoothing_obj$smoothing_type == "independent") {
-    if (is.null(pathogen_names)) {
-      stop("pathogen_names is required for 'independent' smoothing structure")
-    }
+#' @noRd
+validate_smoothing_independent <- function(smoothing_obj, pathogen_names) {
+  if (smoothing_obj$smoothing_type != "independent") return(smoothing_obj)
 
-    expected_dim <- length(pathogen_names)
-
-    # If no priors were provided (priors_provided = 1), create dummy values with correct dimensions
-    if (smoothing_obj$priors_provided == 1) {
-      smoothing_obj$tau_priors$mean <- rep(0.0, expected_dim)
-      smoothing_obj$tau_priors$sd <- rep(1.0, expected_dim)
-    } else {
-      # If priors were provided, validate and adjust dimensions
-      # Validate and adjust dimensions for mean
-      mean_dim <- length(smoothing_obj$tau_priors$mean)
-      if (mean_dim == 1 && expected_dim > 1) {
-        smoothing_obj$tau_priors$mean <- rep(smoothing_obj$tau_priors$mean, expected_dim)
-      } else if (mean_dim != expected_dim && mean_dim > 1) {
-        stop(sprintf(
-          "Length of tau_mean (%d) does not match number of pathogens (%d)",
-          mean_dim, expected_dim
-        ))
-      }
-
-      # Validate and adjust dimensions for sd
-      sd_dim <- length(smoothing_obj$tau_priors$sd)
-      if (sd_dim == 1 && expected_dim > 1) {
-        smoothing_obj$tau_priors$sd <- rep(smoothing_obj$tau_priors$sd, expected_dim)
[…]
