### 2026-03-17 a36e00e
set up to clear cache, update checkout action version
