### 2026-03-17 eb98c17
remove unecessary step to explicitly install piggyback
