+++ .github/workflows/pkgcheck.yaml
@@ -10,7 +10,7 @@ on:
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
-    timeout-minutes: 60
+    timeout-minutes: 360
     permissions:
       issues: write
     steps:
