### 2026-03-18 072e3aa
refresh to simple pkgcheck
