+++ .github/workflows/pkgcheck.yaml
@@ -1,28 +1,22 @@
 name: pkgcheck
+
+# This will cancel running jobs once a new run is triggered
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref }}
   cancel-in-progress: true
+
 on:
+  # Manually trigger the Action under Actions/pkgcheck
   workflow_dispatch:
+  # Run on every push to main
   push:
     branches:
       - main
+
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
-    timeout-minutes: 360
     permissions:
       issues: write
     steps:
-      - uses: actions/checkout@v5
-      - name: Free Disk Space
-        run: |
-          sudo rm -rf /usr/share/dotnet
-          sudo rm -rf /opt/ghc
-          sudo rm -rf /usr/local/share/boost
-          sudo rm -rf "$AGENT_TOOLSDIRECTORY"
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: '4.5.2'
-          use-public-rspm: true
       - uses: ropensci-review-tools/pkgcheck-action@main
