### 2026-03-23 aed45b6
ensure github pat available for fixture downloads
