+++ .github/workflows/pkgcheck.yaml
@@ -1,22 +1,23 @@
 name: pkgcheck
-
-# This will cancel running jobs once a new run is triggered
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref }}
   cancel-in-progress: true
-
 on:
-  # Manually trigger the Action under Actions/pkgcheck
   workflow_dispatch:
-  # Run on every push to main
   push:
     branches:
       - main
-
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
+    timeout-minutes: 360
     permissions:
       issues: write
+      contents: read
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
+      - uses: actions/checkout@v4
       - uses: ropensci-review-tools/pkgcheck-action@main
+        env:
+          GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
