+++ R/fit_model.R
@@ -5,17 +5,26 @@
 #' @param constructed_model prepared model object of class
 #'  `EpiStrainDynamics.model`
 #' @param n_chain number of MCMC chains, defaults to 4
-#' @param n_iter A positive integer specifying the number of iterations for each chain, default value is  2000
-#' @param n_warmup A positive integer specifying the number of warmup iterations, default value is half the number of iterations
-#' @param thin A positive integer specifying the period for saving samples, default value is 1.
-#' @param adapt_delta Numeric value between 0 and 1 indicating target average acceptance probability used in `rstan::sampling`. Default value is 0.9.
-#' @param multi_cores A logical value indicating whether to parallelize chains with multiple cores, default is TRUE and uses all available cores - 1.
-#' @param verbose Logical value controlling the verbosity of output. When TRUE (default),
-#'   shows all messages, warnings, errors, and progress indicators. When FALSE, suppresses
-#'   messages and progress while retaining warnings and errors.
-#' @param suppress_warnings Logical value indicating whether to suppress warnings from Stan.
-#'   Default is FALSE. When TRUE, warnings are suppressed but errors are still raised.
-#' @param seed A positive integer seed used for random number generation in MCMC. Default is NULL, which means the seed is generated from 1 to the maximum integer supported by R.
+#' @param n_iter A positive integer specifying the number of iterations for each
+#'   chain, default value is 2000
+#' @param n_warmup A positive integer specifying the number of warmup
+#'   iterations,default value is half the number of iterations
+#' @param thin A positive integer specifying the period for saving samples,
+#'   default value is 1.
+#' @param adapt_delta Numeric value between 0 and 1 indicating target average
+#'   acceptance probability used in `rstan::sampling`. Default value is 0.9.
+#' @param multi_cores A logical value indicating whether to parallelize chains
+#'   with multiple cores, default is TRUE and uses all available cores - 1.
+#' @param verbose Logical value controlling the verbosity of output. When TRUE
+#'   (default), shows all messages, warnings, errors, and progress indicators.
+#'   When FALSE, suppresses messages and progress while retaining warnings and
+#'   errors.
+#' @param suppress_warnings Logical value indicating whether to suppress
+#'   warnings from Stan. Default is FALSE. When TRUE, warnings are suppressed
+#'   but errors are still raised.
+#' @param seed A positive integer seed used for random number generation in
+#'   MCMC. Default is NULL, which means the seed is generated from 1 to the
+#'   maximum integer supported by R.
 #' @param ... additional arguments to `rstan::sampling()`, such as `init`
 #'
 #' @returns fit model of class `EpiStrainDynamics.fit`, or if fitting fails,
