### 2026-03-23 f09a92b
tidy helper for use only in extended tests, test-fit_model should use only synthetic to run quickly
