+++ tests/testthat/helper-fixtures.R
@@ -1,6 +1,7 @@
+# tests/testthat/helper-fixtures.R
 # Downloads pre-fitted model fixtures from GitHub Release via piggyback.
-# Files that use these: test-fit-model.R, test-diagnose-model.R,
-# test-metrics.R, test-plot.R, test-extended-regression.R
+# Files that use these: test-diagnose-model.R, test-metrics.R,
+# test-plot.R, test-extended-regression.R
 
 skip_if_no_extended_fixtures <- function() {
   skip_if_not(
@@ -9,24 +10,25 @@ skip_if_no_extended_fixtures <- function() {
   )
 }
 
+if (!requireNamespace("piggyback", quietly = TRUE)) {
+  skip("Package 'piggyback' is required to download test fixtures. Install with install.packages('piggyback').")
+}
+
+# Bridge GITHUB_TOKEN to GITHUB_PAT for piggyback authentication
+if (nchar(Sys.getenv("GITHUB_TOKEN")) > 0 && nchar(Sys.getenv("GITHUB_PAT")) == 0) {
+  Sys.setenv(GITHUB_PAT = Sys.getenv("GITHUB_TOKEN"))
+}
+
 message("EpiStrainDynamics: downloading test fixtures via piggyback...")
 
 fixture_dir <- file.path(tempdir(), "testfixtures")
 dir.create(fixture_dir, showWarnings = FALSE)
 
-if (!requireNamespace("piggyback", quietly = TRUE)) {
-  skip("Package 'piggyback' is required to download test fixtures. Install
-       with install.packages('piggyback').")
-}
-# Bridge GITHUB_TOKEN to GITHUB_PAT for piggyback authentication
-if (nchar(Sys.getenv("GITHUB_TOKEN")) > 0 && nchar(Sys.getenv("GITHUB_PAT")) == 0) {
-  Sys.setenv(GITHUB_PAT = Sys.getenv("GITHUB_TOKEN"))
-}
 download_ok <- tryCatch({
   piggyback::pb_download(
-    repo = "acefa-hubs/epistraindynamics",
-    tag = "test-fixtures-v1",
-    dest = fixture_dir,
+    repo      = "acefa-hubs/epistraindynamics",
+    tag       = "test-fixtures-v1",
+    dest      = fixture_dir,
     overwrite = TRUE
   )
   TRUE
@@ -36,40 +38,33 @@ download_ok <- tryCatch({
 })
 
 if (!download_ok) {
-  skip("Fixture download from GitHub Release failed. Check network access and
-       GitHub PAT.")
+  skip("Fixture download from GitHub Release failed. Check network access and GitHub PAT.")
 }
 
 # Regular fixtures (small data subsets, minimal MCMC settings)
-fit_rw_single <<- readRDS(file.path(fixture_dir, "fit_rw_single.rds"))
-fit_ps_single <<- readRDS(file.path(fixture_dir, "fit_ps_single.rds"))
-fit_rw_multi <<- readRDS(file.path(fixture_dir, "fit_rw_multi.rds"))
-fit_ps_multi <<- readRDS(file.path(fixture_dir, "fit_ps_multi.rds"))
-fit_rw_subtyped <<- readRDS(file.path(fixture_dir, "fit_rw_subtyped.rds"))
-fit_ps_subtyped <<- readRDS(file.path(fixture_dir, "fit_ps_subtyped.rds"))
+fit_rw_single     <<- readRDS(file.path(fixture_dir, "fit_rw_single.rds"))
+fit_ps_single     <<- readRDS(file.path(fixture_dir, "fit_ps_single.rds"))
[…]
