+++ tests/testthat/test-fit_model.R
@@ -3,11 +3,46 @@
 # Note: Detailed metric calculations are tested in test-metrics.R
 # Note: Verbosity and error handling (BS2.12-BS2.15) tested in test-verbosity.R
 #' @srrstats {G5.4, G5.5} Correctness tests with fixed test data + fixed seed
-#' @srrstats {G5.11, G5.11a} Extended fixtures downloaded by piggybank. Tests
-#'   skip if download fails.
 
 # Prevent browser from opening during tests (rstan diagnostics)
 options(browser = function(...) {})
+
+# ==============================================================================
+# SETUP: Use lightweight synthetic data
+# ==============================================================================
+
+create_test_data <- function(n_days = 30, seed = 123) {
+  set.seed(seed)
+  data.frame(
+    dates = seq.Date(as.Date("2024-01-01"), by = "day", length.out = n_days),
+    cases = rpois(n_days, lambda = exp(6 + rnorm(n_days, 0, 0.2)))
+  )
+}
+
+create_test_data_multi <- function(n_days = 30, seed = 123) {
+  set.seed(seed)
+  total <- rpois(n_days, lambda = exp(6 + rnorm(n_days, 0, 0.2)))
+
+  # Simple multinomial split
+  props <- matrix(c(0.4, 0.3, 0.2, 0.1), nrow = n_days, ncol = 4, byrow = TRUE)
+  props <- props + matrix(rnorm(n_days * 4, 0, 0.05), nrow = n_days)
+  props <- pmax(props, 0.01)
+  props <- props / rowSums(props)
+
+  p1 <- rbinom(n_days, total, props[, 1])
+  p2 <- rbinom(n_days, total - p1, props[, 2] / (1 - props[, 1]))
+  p3 <- rbinom(n_days, total - p1 - p2, props[, 3] / (1 - props[, 1] - props[, 2]))
+  p4 <- total - p1 - p2 - p3
+
+  data.frame(
+    dates = seq.Date(as.Date("2024-01-01"), by = "day", length.out = n_days),
+    total = total,
+    var1 = p1,
+    var2 = p2,
+    var3 = p3,
+    var4 = p4
+  )
+}
 
 # ==============================================================================
 # TESTS: INPUT VALIDATION
@@ -26,13 +61,13 @@ test_that("fit_model() validates constructed_model class", {
 })
 
 test_that("fit_model() validates MCMC parameters", {
-
+  test_data <- create_test_data(n_days = 30)
   mod <- construct_model(
     method = random_walk(),
     pathogen_structure = single(
-      data = sarscov2,
+      data = test_data,
       case_timeseries = 'cases',
-      time = 'date'
+      time = 'dates'
     )
   )
 
@@ -47,17 +82,28 @@ test_that("fit_model() validates MCMC parameters", {
     fit_model(mod, n_iter = 0),
     "must be positive|invalid"
   )
+
 })
 
 # ==============================================================================
-# TESTS: OUTPUT STRUCTURE (using cached fitted models from setup)
+# TESTS: OUTPUT STRUCTURE (using fast, small models)
 # ==============================================================================
 
 test_that("fit_model() returns correct structure for single pathogen RW", {
-  skip_if_not(exists("fit_rw_single"), "Cached fitted models not available")
+  test_data <- create_test_data(n_days = 30)
+  mod <- construct_model(
+    method = random_walk(),
+    pathogen_structure = single(
+      data = test_data,
+      case_timeseries = 'cases',
+      time = 'dates'
+    )
+  )
 
-  # Use the pre-fitted model from setup
-  fit <- fit_rw_single
+  suppressWarnings(
[…]
