+++ articles/algorithmic-scaling.md
@@ -290,10 +290,10 @@ knitr::kable(scaling_results, digits = 3,
 
 | method      | pathogen_structure | scaling_exponent | r_squared |
 |:------------|:-------------------|-----------------:|----------:|
-| p_spline    | multiple           |            1.019 |     0.997 |
-| p_spline    | single             |            0.846 |     0.991 |
-| random_walk | multiple           |            0.966 |     0.999 |
-| random_walk | single             |            0.878 |     0.982 |
+| p_spline    | multiple           |            1.013 |     0.996 |
+| p_spline    | single             |            0.854 |     0.991 |
+| random_walk | multiple           |            0.967 |     0.999 |
+| random_walk | single             |            0.880 |     0.981 |
 
 Scaling exponents: time complexity approximately O(n^exponent)
 
@@ -787,7 +787,7 @@ on data similar to its intended use case.
 sessionInfo()
 #> R version 4.5.3 (2026-03-11)
 #> Platform: x86_64-pc-linux-gnu
-#> Running under: Ubuntu 24.04.3 LTS
+#> Running under: Ubuntu 24.04.4 LTS
 #> 
 #> Matrix products: default
 #> BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3 
@@ -824,7 +824,7 @@ sessionInfo()
 #> [34] xts_0.14.2            future_1.70.0         inline_0.3.21        
 #> [37] globals_0.19.1        scales_1.4.0          MASS_7.3-65          
 #> [40] anytime_0.3.12        cli_3.6.5             rmarkdown_2.30       
-#> [43] ragg_1.5.1            generics_0.1.4        otel_0.2.0           
+#> [43] ragg_1.5.2            generics_0.1.4        otel_0.2.0           
 #> [46] RcppParallel_5.1.11-2 future.apply_1.20.2   cachem_1.1.0         
 #> [49] rstan_2.32.7          splines_4.5.3         bayesplot_1.15.0     
 #> [52] parallel_4.5.3        matrixStats_1.5.0     vctrs_0.7.2          
@@ -839,6 +839,6 @@ sessionInfo()
 #> [79] textshaping_1.0.5     evaluate_1.0.5        lattice_0.22-9       
 #> [82] tsibble_1.2.0         bslib_0.10.0          rstantools_2.6.0     
 #> [85] class_7.3-23          Rcpp_1.1.1            gridExtra_2.3        
-#> [88] prodlim_2026.03.11    xfun_0.57             fs_2.0.0             
+#> [88] prodlim_2026.03.11    xfun_0.57             fs_2.0.1             
 #> [91] zoo_1.8-15            pkgconfig_2.0.3
 ```
