### 2026-05-08 109d66a
add ropensci badge to readme
