### 2026-05-08 99ce78f
begin a news file to capture changes made during review
