+++ NEWS.md
@@ -1,0 +1,7 @@
+
+EpiStrainDynamics 0.0.1.0000 (2026-03-27)
+=========================
+
+### New features
+
+  * submitted to ROpenSci
