+++ NAMESPACE
@@ -38,13 +38,27 @@ export(single)
 export(smoothing_structure)
 export(subtyped)
 import(Rcpp)
-import(ggplot2)
 import(methods)
 importFrom(RcppParallel,RcppParallelLibs)
 importFrom(bayesplot,mcmc_areas)
 importFrom(bayesplot,mcmc_trace)
 importFrom(dplyr,arrange)
 importFrom(dplyr,bind_rows)
+importFrom(ggplot2,aes)
+importFrom(ggplot2,element_blank)
+importFrom(ggplot2,geom_hline)
+importFrom(ggplot2,geom_line)
+importFrom(ggplot2,geom_point)
+importFrom(ggplot2,geom_ribbon)
+importFrom(ggplot2,ggplot)
+importFrom(ggplot2,ggplot_build)
+importFrom(ggplot2,scale_colour_manual)
+importFrom(ggplot2,scale_y_continuous)
+importFrom(ggplot2,sec_axis)
+importFrom(ggplot2,theme)
+importFrom(ggplot2,theme_bw)
+importFrom(ggplot2,xlab)
+importFrom(ggplot2,ylab)
 importFrom(purrr,map2)
 importFrom(rlang,.data)
 importFrom(rstan,extract)
