+++ articles/algorithmic-scaling.md
@@ -22,6 +22,7 @@ We simulate realistic epidemic dynamics using SIR models with pathogen
 succession patterns.
 
 ``` r
+
 # SIR model simulation function
 simulate_sir <- function(n_days, R0, recovery_rate = 0.1, I0 = 0.01) {
   S <- numeric(n_days)
@@ -109,6 +110,7 @@ data_sizes <- c(30, 60, 120, 240, 480)
 ### 1.2 Benchmark Random Walk Method
 
 ``` r
+
 rw_timings <- data.frame(
   n_timepoints = integer(),
   method = character(),
@@ -179,6 +181,7 @@ for (n in data_sizes) {
 ### 1.3 Benchmark P-Spline Method
 
 ``` r
+
 ps_timings <- data.frame(
   n_timepoints = integer(),
   method = character(),
@@ -247,6 +250,7 @@ for (n in data_sizes) {
 ### 1.4 Visualize Scaling Behavior
 
 ``` r
+
 # Combine all timings
 all_timings <- rbind(rw_timings, ps_timings)
 
@@ -274,6 +278,7 @@ ggplot(all_timings, aes(x = n_timepoints, y = time_seconds,
 ### 1.5 Quantify Scaling Coefficients
 
 ``` r
+
 # Fit linear models on log-log scale to estimate scaling exponent
 scaling_results <- all_timings %>%
   group_by(method, pathogen_structure) %>%
@@ -290,12 +295,12 @@ knitr::kable(scaling_results, digits = 3,
 
 | method      | pathogen_structure | scaling_exponent | r_squared |
 |:------------|:-------------------|-----------------:|----------:|
-| p_spline    | multiple           |            1.013 |     0.996 |
-| p_spline    | single             |            0.854 |     0.991 |
-| random_walk | multiple           |            0.967 |     0.999 |
-| random_walk | single             |            0.880 |     0.981 |
+| p_spline    | multiple           |            1.014 |     0.996 |
+| p_spline    | single             |            0.849 |     0.992 |
+| random_walk | multiple           |            0.957 |     0.999 |
+| random_walk | single             |            0.881 |     0.982 |
 
-Scaling exponents: time complexity approximately O(n^exponent)
+Scaling exponents: time complexity approximately O(n^exponent) {.table}
 
 **Interpretation**: Exponents close to 1.0 indicate linear scaling O(n),
 which is expected for these Bayesian time series models. Values slightly
@@ -310,6 +315,7 @@ We verify that fitted/predicted values are on the same scale as the
 input data.
 
 ``` r
+
 # Use package data
 data(sarscov2)
 
@@ -363,9 +369,10 @@ knitr::kable(input_summary, digits = 2,
 | max    |   275647.00 |       266442.25 |  0.97 |
 | sd     |    34268.32 |        33883.50 |  0.99 |
 
-Comparison of input and predicted case scales
+Comparison of input and predicted case scales {.table}
 
 ``` r
+
 # Visual comparison with credible intervals
 plot_data <- data.frame(
   date = sarscov2$date,
@@ -408,6 +415,7 @@ Test whether the model correctly recovers input scales when data have
 different properties (non-zero means, different magnitudes, etc.).
 
 ``` r
+
 # Test with data at different scales 
 test_scales <- c(10, 100, 1000)
 recovery_results <- data.frame(
@@ -455,9 +463,10 @@ knitr::kable(recovery_results, digits = 2,
[…]
