+++ articles/parameter-recovery.md
@@ -34,6 +34,7 @@ dynamics.
 ### 1.1 Generate Data with Known SIR Dynamics
 
 ``` r
+
 set.seed(42)
 
 # SIR model simulation
@@ -92,6 +93,7 @@ true_values_single <- data.frame(
 ### 1.2 Fit Model with Multiple Seeds
 
 ``` r
+
 seeds <- c(111, 222)
 single_fits <- list()
 
@@ -129,6 +131,7 @@ for (i in seq_along(seeds)) {
 ### 1.3 Evaluate Recovery
 
 ``` r
+
 # Calculate metrics for each seed
 recovery_metrics_single <- data.frame(
   seed = seeds,
@@ -155,9 +158,10 @@ knitr::kable(
 |  111 |      0.9918 | 5.5725 |        0.1120 |           1 |      0.3444 |
 |  222 |      0.9918 | 5.5910 |        0.1124 |           1 |      0.3556 |
 
-Single pathogen parameter recovery metrics across seeds
+Single pathogen parameter recovery metrics across seeds {.table}
 
 ``` r
+
 
 cat("\nSummary across seeds:\n")
 #> 
@@ -173,6 +177,7 @@ cat("Mean 50% CI coverage:", round(mean(recovery_metrics_single$coverage_50), 4)
 ```
 
 ``` r
+
 # Plot first seed results
 plot_data <- data.frame(
   date = dates,
@@ -217,6 +222,7 @@ succession dynamics.
 ### 2.1 Generate Data with Known Pathogen Succession
 
 ``` r
+
 set.seed(123)
 
 n_timepoints <- 120
@@ -265,6 +271,7 @@ multiple_data <- data.frame(
 ### 2.2 Fit Model with Multiple Seeds
 
 ``` r
+
 seeds <- c(555, 666)
 multiple_fits <- list()
 
@@ -298,6 +305,7 @@ for (i in seq_along(seeds)) {
 ### 2.3 Evaluate Proportion Recovery
 
 ``` r
+
 # Calculate recovery metrics for each pathogen and seed
 recovery_metrics_multiple <- data.frame()
 
@@ -340,11 +348,12 @@ knitr::kable(summary_by_pathogen, digits = 4,
 | pathogen3 |           0.9980 |    0.0256 |   0.0132 |         0.0009 |
 
 Multiple pathogen proportion recovery summary across seeds (G5.6a,
-G5.6b)
+G5.6b) {.table}
 
 ### 2.4 Visualize True vs Estimated Proportions
 
 ``` r
+
 # Plot recovery for first seed with true proportions
 props_first <- multiple_fits[[1]]$proportions
 prop1 <- props_first$measure$y[props_first$measure$pathogen == 'pathogen1']
@@ -380,6 +389,7 @@ ggplot(plot_data, aes(x = date)) +
 ![](parameter-recovery_files/figure-html/plot-multiple-recovery-1.png)
 
 ``` r
+
 # Create comparison plots showing both true and estimated side by side
 comparison_data <- data.frame(
   date = rep(dates, 6),
@@ -412,6 +422,7 @@ ggplot(comparison_data, aes(x = date, y = proportion, color = type, linetype = t
 ### 2.5 Consistency Across Seeds (G5.6b)
 
 ``` r
+
 # Compare estimates across seeds at specific time points
 check_times <- c(30, 60, 90)
 
@@ -449,7 +460,7 @@ knitr::kable(consistency_results, digits = 4,
 | pathogen31 | pathogen3 |        60 |     0.0000 |        0.0157 |      0.0003 | 0.0167 |
 | pathogen32 | pathogen3 |        90 |     0.8209 |        0.8418 |      0.0581 | 0.0690 |
 
-Consistency of estimates across random seeds at key timepoints
[…]
