### 2026-05-08 9ee2040
replace slow examples that used \dontrun with @examplesIf interactive().
where dontrun was used because of package dependency, use @examplesIf interactive()

closes #36
