### 2026-05-08 f44b4ee
ensure arguments with default are listed after those without.
closes #38
