### 2026-05-08 f4478ed
closes #39
