+++ DESCRIPTION
@@ -59,5 +59,6 @@ Suggests:
     xts,
     zoo
 Config/testthat/edition: 3
+Config/testthat/parallel: true
 LazyData: true
 VignetteBuilder: knitr
