### 2026-05-18 1466c44
needed to move values to lower case before doing arg_match
