+++ tests/testthat/test-validation-utils.R
@@ -2,14 +2,16 @@
 #'   including conditions that trigger these messages and comparing with
 #'   expected results
 #' @srrstats {G3.0} comparisons made between appropriate values
-#'
-# Tests for validate_positive_whole_number()
+
+# ==============================================================================
+# TESTS: validate_positive_whole_number()
+# ==============================================================================
+
 test_that("validate_positive_whole_number() passes valid inputs", {
-  # Should not throw errors for valid positive whole numbers
   expect_silent(EpiStrainDynamics:::validate_positive_whole_number(1, "test_param"))
   expect_silent(EpiStrainDynamics:::validate_positive_whole_number(5L, "test_param"))
   expect_silent(EpiStrainDynamics:::validate_positive_whole_number(100, "test_param"))
-  expect_silent(EpiStrainDynamics:::validate_positive_whole_number(3.0, "test_param"))  # Whole number as double
+  expect_silent(EpiStrainDynamics:::validate_positive_whole_number(3.0, "test_param"))
 })
 
 test_that("validate_positive_whole_number() returns invisible NULL", {
@@ -72,7 +74,10 @@ test_that("validate_positive_whole_number() uses correct argument name in errors
                "Argument days_per_knot must be numeric")
 })
 
-# Tests for validate_class_inherits()
+# ==============================================================================
+# TESTS: validate_class_inherits()
+# ==============================================================================
+
 test_that("validate_class_inherits() passes valid class", {
   df <- data.frame(x = 1:3, y = letters[1:3])
   mat <- matrix(1:6, nrow = 2)
@@ -83,64 +88,11 @@ test_that("validate_class_inherits() passes valid class", {
 })
 
 test_that("validate_class_inherits() handles inheritance", {
-  # Create object with multiple classes
   obj <- structure(list(a = 1), class = c("child_class", "parent_class"))
-
   expect_silent(EpiStrainDynamics:::validate_class_inherits(obj, "parent_class"))
   expect_silent(EpiStrainDynamics:::validate_class_inherits(obj, "child_class"))
 })
 
-# Tests for validate_gi_dist()
-test_that("validate_gi_dist accepts valid functions", {
-  # Standard valid function
-  gi_dist <- function(x) 4 * x * exp(-2 * x)
-  expect_silent(validate_gi_dist(gi_dist))
-
-  # Exponential-like
-  gi_dist1 <- function(x) exp(-x)
-  expect_silent(validate_gi_dist(gi_dist1))
-
-  # Gamma-like
-  gi_dist2 <- function(x) x^2 * exp(-x)
-  expect_silent(validate_gi_dist(gi_dist2))
-
-  # Constant
-  gi_dist3 <- function(x) rep(0.1, length(x))
-  expect_silent(validate_gi_dist(gi_dist3))
-})
-
-test_that("validate_gi_dist rejects invalid input types", {
-  # Non-function inputs
-  expect_error(validate_gi_dist(5), class = "rlang_error")
-  expect_error(validate_gi_dist("not a function"), class = "rlang_error")
-
-  # Function with no arguments
-  gi_dist <- function() 1
-  expect_error(validate_gi_dist(gi_dist), class = "rlang_error")
-})
-
-test_that("validate_gi_dist rejects functions with invalid outputs", {
-  # Function that errors
-  gi_dist1 <- function(x) stop("intentional error")
-  expect_error(validate_gi_dist(gi_dist1), class = "rlang_error")
-
-  # Non-numeric output
-  gi_dist2 <- function(x) as.character(x)
-  expect_error(validate_gi_dist(gi_dist2), class = "rlang_error")
-
[…]
