### 2026-05-18 6c78877
add tests for mcmc validation functions and smoothing structure functions
