+++ tests/testthat/test-validation-smoothing.R
@@ -1,0 +1,225 @@
+#' @srrstats {G5.2, G5.2b} error and warning behaviour explicitly tested,
+#'   including conditions that trigger these messages and comparing with
+#'   expected results
+#' @srrstats {G3.0} comparisons made between appropriate values
+
+# ==============================================================================
+# TESTS: smoothing_structure() constructor
+# ==============================================================================
+
+test_that("smoothing_structure() returns correct class", {
+  result <- smoothing_structure("shared")
+  expect_s3_class(result, "EpiStrainDynamics.smoothing")
+})
+
+test_that("smoothing_structure() accepts all valid types", {
+  expect_silent(smoothing_structure("shared"))
+  expect_silent(smoothing_structure("independent"))
+  expect_silent(smoothing_structure("correlated"))
+})
+
+test_that("smoothing_structure() is case insensitive", {
+  expect_silent(smoothing_structure("SHARED"))
+  expect_silent(smoothing_structure("Independent"))
+  expect_silent(smoothing_structure("CORRELATED"))
+})
+
+test_that("smoothing_structure() rejects invalid type", {
+  expect_error(smoothing_structure("invalid"), class = "rlang_error")
+})
+
+test_that("smoothing_structure() alerts when tau priors provided for correlated type", {
+  expect_message(
+    smoothing_structure("correlated", tau_mean = 0.5, tau_sd = 1.0),
+    "ignored for"
+  )
+})
+
+test_that("smoothing_structure() accepts priors for shared type", {
+  result <- smoothing_structure("shared", tau_mean = 0.5, tau_sd = 1.0)
+  expect_s3_class(result, "EpiStrainDynamics.smoothing")
+  expect_equal(result$smoothing_type, "shared")
+  expect_equal(result$priors_provided, 2)
+})
+
+test_that("smoothing_structure() accepts priors for independent type", {
+  result <- smoothing_structure("independent",
+                                tau_mean = c(0.1, 0.2),
+                                tau_sd = c(1.0, 1.0))
+  expect_s3_class(result, "EpiStrainDynamics.smoothing")
+  expect_equal(result$smoothing_type, "independent")
+  expect_equal(result$priors_provided, 2)
+})
+
+test_that("smoothing_structure() sets priors_provided = 1 when no priors given", {
+  result <- smoothing_structure("shared")
+  expect_equal(result$priors_provided, 1)
+})
+
+# ==============================================================================
+# TESTS: validate_smoothing_structure()
+# ==============================================================================
+
+test_that("validate_smoothing_structure() errors on wrong class", {
+  bad_obj <- list(smoothing_type = "shared", tau_priors = list(), priors_provided = 1)
+  expect_error(
+    EpiStrainDynamics:::validate_smoothing_structure(bad_obj, "single"),
+    "smoothing_params must be created using the smoothing_structure\\(\\) function"
+  )
+})
+
+test_that("validate_smoothing_structure() returns smoothing object", {
+  obj <- smoothing_structure("shared")
+  result <- EpiStrainDynamics:::validate_smoothing_structure(obj, "single")
+  expect_s3_class(result, "EpiStrainDynamics.smoothing")
+})
+
+# ==============================================================================
+# TESTS: validate_smoothing_for_single()
+# ==============================================================================
+
+test_that("validate_smoothing_for_single() passes through non-single pathogen", {
+  obj <- smoothing_structure("independent")
[…]
