+++ tests/testthat/test-validation-mcmc.R
@@ -1,0 +1,437 @@
+#' @srrstats {G5.2, G5.2b} error and warning behaviour explicitly tested,
+#'   including conditions that trigger these messages and comparing with
+#'   expected results
+#' @srrstats {G3.0} comparisons made between appropriate values
+
+# ==============================================================================
+# TESTS: validate_n_chain()
+# ==============================================================================
+
+test_that("validate_n_chain() passes valid inputs", {
+  expect_silent(EpiStrainDynamics:::validate_n_chain(1))
+  expect_silent(EpiStrainDynamics:::validate_n_chain(4))
+  expect_silent(EpiStrainDynamics:::validate_n_chain(1L))
+})
+
+test_that("validate_n_chain() rejects non-numeric", {
+  expect_error(EpiStrainDynamics:::validate_n_chain("4"), "must be numeric")
+  expect_error(EpiStrainDynamics:::validate_n_chain(TRUE), "must be numeric")
+})
+
+test_that("validate_n_chain() rejects multiple values", {
+  expect_error(EpiStrainDynamics:::validate_n_chain(c(1, 2)), "must be a single value")
+})
+
+test_that("validate_n_chain() rejects non-finite values", {
+  expect_error(EpiStrainDynamics:::validate_n_chain(Inf), "must be a finite number")
+  expect_error(EpiStrainDynamics:::validate_n_chain(NA_real_), "must be a finite number")
+})
+
+test_that("validate_n_chain() rejects non-whole numbers", {
+  expect_error(EpiStrainDynamics:::validate_n_chain(1.5), "must be a whole number")
+})
+
+test_that("validate_n_chain() rejects non-positive values", {
+  expect_error(EpiStrainDynamics:::validate_n_chain(0), "must be positive")
+  expect_error(EpiStrainDynamics:::validate_n_chain(-1), "must be positive")
+})
+
+# ==============================================================================
+# TESTS: validate_n_iter()
+# ==============================================================================
+
+test_that("validate_n_iter() passes valid inputs", {
+  expect_silent(EpiStrainDynamics:::validate_n_iter(100))
+  expect_silent(EpiStrainDynamics:::validate_n_iter(2000L))
+})
+
+test_that("validate_n_iter() rejects non-numeric", {
+  expect_error(EpiStrainDynamics:::validate_n_iter("100"), "must be numeric")
+})
+
+test_that("validate_n_iter() rejects multiple values", {
+  expect_error(EpiStrainDynamics:::validate_n_iter(c(100, 200)), "must be a single value")
+})
+
+test_that("validate_n_iter() rejects non-finite values", {
+  expect_error(EpiStrainDynamics:::validate_n_iter(Inf), "must be a finite number")
+  expect_error(EpiStrainDynamics:::validate_n_iter(NA_real_), "must be a finite number")
+})
+
+test_that("validate_n_iter() rejects non-whole numbers", {
+  expect_error(EpiStrainDynamics:::validate_n_iter(100.5), "must be a whole number")
+})
+
+test_that("validate_n_iter() rejects non-positive values", {
+  expect_error(EpiStrainDynamics:::validate_n_iter(0), "must be positive")
+  expect_error(EpiStrainDynamics:::validate_n_iter(-100), "must be positive")
+})
+
+# ==============================================================================
+# TESTS: validate_n_warmup()
+# ==============================================================================
+
+test_that("validate_n_warmup() passes valid inputs", {
+  expect_silent(EpiStrainDynamics:::validate_n_warmup(0))
[…]
