### 2026-05-18 577a42f
incorporate test for incorrectly adding tau arguments for "correlated" structure
