### 2026-05-18 eaace05
split validation-utils into three files so it is easier to examine closely
