+++ R/validation-mcmc.R
@@ -1,0 +1,502 @@
+#' Validate R-hat Threshold Parameter
+#'
+#' Validates that rhat_threshold is a positive numeric value greater than or equal to 1
+#'
+#' @param rhat_threshold The R-hat convergence threshold
+#'
+#' @noRd
+#' @srrstats {G1.4} uses `Roxygen2` documentation
+#' @srrstats {G1.4a} internal function specified with `@noRd`
+#' @srrstats {G5.2a} every error statement is unique
+#'
+#' @return NULL (invisibly) if validation passes, otherwise stops with error
+validate_rhat_threshold <- function(rhat_threshold) {
+  if (!is.numeric(rhat_threshold)) {
+    cli::cli_abort("{.arg rhat_threshold} must be numeric, got {.cls {class(rhat_threshold)}}")
+  }
+
+  if (length(rhat_threshold) != 1) {
+    cli::cli_abort("{.arg rhat_threshold} must be a single value, got length {length(rhat_threshold)}")
+  }
+
+  if (!is.finite(rhat_threshold)) {
+    cli::cli_abort("{.arg rhat_threshold} must be a finite number")
+  }
+
+  if (rhat_threshold < 1.0) {
+    cli::cli_abort(c(
+      "{.arg rhat_threshold} must be greater than or equal to 1.0",
+      "x" = "R-hat values are always >= 1 by definition",
+      "i" = "Common thresholds: 1.01 (strict), 1.05 (moderate), 1.1 (lenient)"
+    ))
+  }
+
+  # Warn if threshold seems unusual
+  if (rhat_threshold > 2.0) {
+    cli::cli_warn(c(
+      "{.arg rhat_threshold} = {rhat_threshold} is unusually high",
+      "i" = "This may not detect convergence issues effectively",
+      "i" = "Consider using a value between 1.01 and 1.1"
+    ))
+  }
+
+  invisible(NULL)
+}
+
+#' Validate Effective Sample Size Threshold Parameter
+#'
+#' Validates that eff_sample_threshold is a positive numeric value
+#'
+#' @param eff_sample_threshold The minimum effective sample size threshold
+#'
+#' @noRd
+#' @srrstats {G1.4} uses `Roxygen2` documentation
+#' @srrstats {G1.4a} internal function specified with `@noRd`
+#' @srrstats {G5.2a} every error statement is unique
+#'
+#' @return NULL (invisibly) if validation passes, otherwise stops with error
+validate_eff_sample_threshold <- function(eff_sample_threshold) {
+  if (!is.numeric(eff_sample_threshold)) {
+    cli::cli_abort("{.arg eff_sample_threshold} must be numeric, got {.cls {class(eff_sample_threshold)}}")
+  }
+
+  if (length(eff_sample_threshold) != 1) {
+    cli::cli_abort("{.arg eff_sample_threshold} must be a single value, got length {length(eff_sample_threshold)}")
+  }
+
+  if (!is.finite(eff_sample_threshold)) {
+    cli::cli_abort("{.arg eff_sample_threshold} must be a finite number")
+  }
+
+  if (eff_sample_threshold <= 0) {
+    cli::cli_abort(c(
+      "{.arg eff_sample_threshold} must be positive, got {eff_sample_threshold}",
+      "i" = "Effective sample size represents the number of independent samples",
+      "i" = "Common thresholds: 100 (minimum), 400 (good), 1000 (excellent)"
+    ))
+  }
+
+  # Warn if threshold seems unusual
+  if (eff_sample_threshold < 50) {
+    cli::cli_warn(c(
+      "{.arg eff_sample_threshold} = {eff_sample_threshold} is very low",
[…]

[…]
