### 2026-05-18 3241d7c
add revision notes from editor comments https://github.com/ropensci/software-review/issues/763#issuecomment-4406583890
