+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: EpiStrainDynamics
 Title: Infer temporal trends of multiple pathogens
-Version: 0.0.1.0000
+Version: 0.0.1.1
 Authors@R: 
     c(person("Saras", "Windecker", email = "saras.windecker@gmail.com", 
         role = c("aut", "cre"),
