+++ reference/multiple.md
@@ -5,7 +5,7 @@ Multiple pathogen structure
 ## Usage
 
 ``` r
-multiple(data, case_timeseries, time = NULL, component_pathogen_timeseries)
+multiple(data, case_timeseries, component_pathogen_timeseries, time = NULL)
 ```
 
 ## Arguments
@@ -20,18 +20,18 @@ multiple(data, case_timeseries, time = NULL, component_pathogen_timeseries)
   Column name containing case counts. Must be numeric or a `units`
   object from the units package.
 
-- time:
-
-  name of column with time data. Required for non-time-series input
-  data. Flexible format - can be date, index, or others, accepted as
-  `index` identifiers in the `tsibble` time format. Optional when `data`
-  is a time series class object (ts, mts, xts, zoo, zooreg, tsibble) as
-  the time index will be automatically detected.
-
 - component_pathogen_timeseries:
 
   vector of column names with additional pathogen case count timeseries
   to model. Must be numeric or a `units` object from the units package.
+
+- time:
+
+  Column name with time data. Required for non-time-series input data.
+  Flexible format - can be date, index, or others, accepted as `index`
+  identifiers in the `tsibble` time format. Optional when `data` is a
+  time series class object (ts, mts, xts, zoo, zooreg, tsibble) as the
+  time index will be automatically detected.
 
 ## Value
 
@@ -51,8 +51,8 @@ Other pathogen_structure:
 multiple(
   data = sarscov2,
   case_timeseries = 'cases',
-  time = 'date',
-  component_pathogen_timeseries = c('alpha', 'delta', 'omicron', 'other')
+  component_pathogen_timeseries = c('alpha', 'delta', 'omicron', 'other'),
+  time = 'date'
 )
 #> $pathogen_structure
 #> [1] "multiple"
@@ -580,7 +580,6 @@ multiple(
 #> [2] "list"                                
 
 # Using a time series object (time argument is optional)
-if (FALSE) { # \dontrun{
 sarscov2_xts <- xts::xts(
   sarscov2[, c("cases", "alpha", "delta", "omicron", "other")],
   order.by = sarscov2$date
@@ -590,5 +589,528 @@ multiple(
   case_timeseries = 'cases',
   component_pathogen_timeseries = c('alpha', 'delta', 'omicron', 'other')
 )
-} # }
+#> $pathogen_structure
+#> [1] "multiple"
+#> 
+#> $pathogen_names
+#> [1] "alpha"   "delta"   "omicron" "other"  
+#> 
+#> $validated_tsbl
+#> # A tsibble: 830 x 6 [1D]
+#>    index      case_timeseries alpha delta omicron other
+#>    <date>               <dbl> <dbl> <dbl>   <dbl> <dbl>
+#>  1 2020-09-23            7096     0     0       0    10
+#>  2 2020-09-24            7682     0     0       0     7
+#>  3 2020-09-25            7318     0     0       0     9
+#>  4 2020-09-26            6810     0     0       0    15
+#>  5 2020-09-27            7023     0     0       0     5
+#>  6 2020-09-28            9983     0     0       0     6
+#>  7 2020-09-29           10366     0     0       0    23
+#>  8 2020-09-30           12717     0     0       0    54
+#>  9 2020-10-01           13357     0     0       0    32
+#> 10 2020-10-02           13854     0     0       0    15
+#> # ℹ 820 more rows
+#> 
+#> $data
+#> $data$case_timeseries
+#>   [1]   7096   7682   7318   6810   7023   9983  10366  12717  13357  13854
+#>  [11]  11667  11953  16753  17250  18573  18425  15893  12648  12202  19671
[…]
