+++ reference/plot.md
@@ -47,7 +47,7 @@ ggplot2 plot output
 ## Examples
 
 ``` r
-if (FALSE) { # \dontrun{
+if (FALSE) { # interactive()
   mod <- construct_model(
     method = random_walk(),
     pathogen_structure = single(
@@ -57,5 +57,5 @@ if (FALSE) { # \dontrun{
   fit <- fit_model(mod)
   gr <- growth_rate(mod)
   plot(gr)
-} # }
+}
 ```
