+++ reference/construct_model.md
@@ -37,7 +37,7 @@ construct_model(
   argument is optional and defines the structure of the smoothing terms
   including optionally setting the smoothing prior tau. Created with
   [`smoothing_structure()`](https://acefa-hubs.github.io/EpiStrainDynamics/reference/smoothing_structure.md).
-  NULL option defaults to 'shared' smoothing structure and default
+  NULL option defaults to "shared" smoothing structure and default
   priors.
 
 - dispersion_params:
@@ -73,13 +73,13 @@ mod <- construct_model(
 
   pathogen_structure = multiple(
     data = sarscov2,
-    case_timeseries = 'cases',
-    time = 'date',
-    component_pathogen_timeseries = c('alpha', 'delta', 'omicron', 'other')
+    case_timeseries = "cases",
+    time = "date",
+    component_pathogen_timeseries = c("alpha", "delta", "omicron", "other")
     ),
 
    smoothing_params = smoothing_structure(
-      'independent', tau_mean = c(0, 0.1, 0.3, 0), tau_sd = rep(1, times = 4)),
+      "independent", tau_mean = c(0, 0.1, 0.3, 0), tau_sd = rep(1, times = 4)),
    dispersion_params = dispersion_structure(phi_mean = 0, phi_sd = 1),
    pathogen_noise = FALSE,
    dow_effect = TRUE
